Module Ansi_kernel

Common ANSI display attribute definitions.

NOTE: assorted content lifted from lib/console/src/console.ml

module Color_256 : sig ... end

Support for 256-color handling on terminals/consoles. Note that the functions of_rgb6_exn and of_rgb6 return values within the 6x6x6 color-cube space, even though equivalent duplicates exist in the first 16 and last 24 colors (a "best-fit-equivalent" function could be added some day, if there ever becomes a requirement for it -- see note 2 below).

module Color : sig ... end

module Attr : sig ... end

module With_all_attrs : sig ... end