Parameter Make_and_test.1-T

include Base.Binary_searchable.Indexable

type elt

type t

val get : t -> int -> elt

val length : t -> int

module For_test : Test_binary_searchable_intf.For_test with type elt := elt with type t := t

To implement the tests, we need two different elt values small < big, to be able to compare those values, and to be able to construct a t containing those values.