Module type Test_binary_searchable_intf.Binary_searchable_and_for_test

include Base.Binary_searchable.S

type elt

type t

val binary_search : ?pos:int -> ?len:int -> t -> compare:(elt -> 'key -> int) -> [ `Last_strictly_less_than | `Last_less_than_or_equal_to | `Last_equal_to | `First_equal_to | `First_greater_than_or_equal_to | `First_strictly_greater_than ] -> 'key -> int option

See Binary_search.binary_search in binary_search.ml

val binary_search_segmented : ?pos:int -> ?len:int -> t -> segment_of:(elt -> [ `Left | `Right ]) -> [ `Last_on_left | `First_on_right ] -> int option

See Binary_search.binary_search_segmented in binary_search.ml

module For_test : For_test with type elt := elt with type t := t

To implement the tests, we need two different elt values small < big, to be able to compare those values, and to be able to construct a t containing those values.