Module Core_kernel.Bag

module type S = sig ... end

The module type of the Bag module.

include S

module Elt : sig ... end

type 'a t

include Ppx_sexp_conv_lib.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

Much of a bag's interface comes from the generic Base.Container module.

include Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__Container_intf.Export.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

include Base.Invariant.S1 with type 'a t := 'a t

val invariant : ('a -> unit) -> 'a t -> unit

val create : Base.Unit.t -> 'a t

create () returns an empty bag.

val add : 'a t -> 'a -> 'a Elt.t

add t v adds v to the bag t, returning an element that can later be removed from the bag. add runs in constant time.

val add_unit : 'a t -> 'a -> Base.Unit.t

val mem_elt : 'a t -> 'a Elt.t -> Base.Bool.t

mem_elt t elt returns whether or not elt is in t. It is like mem (included from Container), but it takes an 'a Elt.t instead of an 'a and runs in constant time instead of linear time.

val remove : 'a t -> 'a Elt.t -> Base.Unit.t

remove t elt removes elt from the bag t, raising an exception if elt is not in the bag. remove runs in constant time.

val choose : 'a t -> 'a Elt.t Base.Option.t

choose t returns some element in the bag.

val remove_one : 'a t -> 'a Base.Option.t

remove_one t removes some element from the bag, and returns its value. remove_one runs in constant time.

val clear : 'a t -> Base.Unit.t

clear t removes all elements from the bag. clear runs in constant time.

val filter_inplace : 'a t -> f:('a -> Base.Bool.t) -> Base.Unit.t

filter_inplace t ~f removes all the elements from t that don't satisfy f.

val iter_elt : 'a t -> f:('a Elt.t -> Base.Unit.t) -> Base.Unit.t

val find_elt : 'a t -> f:('a -> Base.Bool.t) -> 'a Elt.t Base.Option.t

find_elt t ~f returns the first element in the bag satisfying f, returning None if none is found.

val until_empty : 'a t -> ('a -> Base.Unit.t) -> Base.Unit.t

until_empty t f repeatedly removes values v from t, running f v on each one, until t is empty. Running f may add elements to t if it wants.

val transfer : src:'a t -> dst:'a t -> Base.Unit.t

transfer ~src ~dst moves all of the elements from src to dst in constant time.

val of_list : 'a Base.List.t -> 'a t

val elts : 'a t -> 'a Elt.t Base.List.t

val unchecked_iter : 'a t -> f:('a -> Base.Unit.t) -> Base.Unit.t

unchecked_iter t ~f behaves like iter t ~f except that f is allowed to modify t. Elements added by f may or may not be visited; elements removed by f that have not been visited will not be visited. It is an (undetected) error to delete the current element.