Module Core_kernel.Bigbuffer

Extensible string buffers based on Bigstrings.

This module implements string buffers that automatically expand as necessary. It provides accumulative concatenation of strings in quasi-linear time (instead of quadratic time when strings are concatenated pairwise).

This implementation uses Bigstrings instead of strings. This removes the 16MB limit on buffer size (on 32-bit machines), and improves I/O-performance when reading/writing from/to channels.

include Base.Buffer.S

type t

The abstract type of buffers.

val sexp_of_t : t -> Base.Sexp.t

val create : int -> t

create n returns a fresh buffer, initially empty. The n parameter is the initial size of the internal storage medium that holds the buffer contents. That storage is automatically reallocated when more than n characters are stored in the buffer, but shrinks back to n characters when reset is called.

For best performance, n should be of the same order of magnitude as the number of characters that are expected to be stored in the buffer (for instance, 80 for a buffer that holds one output line). Nothing bad will happen if the buffer grows beyond that limit, however. In doubt, take n = 16 for instance.

val contents : t -> string

Return a copy of the current contents of the buffer. The buffer itself is unchanged.

val contents_bytes : t -> bytes

include Base.Blit.S_distinct with type src := t with type dst := bytes

val blit : src:t -> src_pos:int -> dst:bytes -> dst_pos:int -> len:int -> unit

val blito : src:t -> ?⁠src_pos:int -> ?⁠src_len:int -> dst:bytes -> ?⁠dst_pos:int -> unit -> unit

val unsafe_blit : src:t -> src_pos:int -> dst:bytes -> dst_pos:int -> len:int -> unit

val sub : t -> pos:int -> len:int -> bytes

val subo : ?⁠pos:int -> ?⁠len:int -> t -> bytes

module To_string : Base.Blit.S_to_string with type t := t

val nth : t -> int -> char

Gets the (zero-based) n-th character of the buffer. Raises Invalid_argument if index out of bounds.

val length : t -> int

Returns the number of characters currently contained in the buffer.

val clear : t -> unit

Empties the buffer.

val reset : t -> unit

Empties the buffer and deallocates the internal storage holding the buffer contents, replacing it with the initial internal storage of length n that was allocated by create n. For long-lived buffers that may have grown a lot, reset allows faster reclamation of the space used by the buffer.

val add_char : t -> char -> unit

add_char b c appends the character c at the end of the buffer b.

val add_string : t -> string -> unit

add_string b s appends the string s at the end of the buffer b.

val add_substring : t -> string -> pos:int -> len:int -> unit

add_substring b s pos len takes len characters from offset pos in string s and appends them at the end of the buffer b.

val add_bytes : t -> bytes -> unit

add_bytes b s appends the bytes s at the end of the buffer b.

val add_subbytes : t -> bytes -> pos:int -> len:int -> unit

add_subbytes b s pos len takes len characters from offset pos in bytes s and appends them at the end of the buffer b.

val add_buffer : t -> t -> unit

add_buffer b1 b2 appends the current contents of buffer b2 at the end of buffer b1. b2 is not modified.

val big_contents : t -> Bigstring.t

Return a copy of the current contents of the buffer as a bigstring. The buffer itself is unchanged.

val volatile_contents : t -> Bigstring.t

Return the actual underlying bigstring used by this bigbuffer. No copying is involved. To be safe, use and finish with the returned value before calling any other function in this module on the same Bigbuffer.t.

val add_bigstring : t -> Bigstring.t -> Base.Unit.t

add_bigstring b s appends the bigstring s at the end of the buffer b.

val add_bin_prot : t -> 'a Bin_prot.Type_class.writer -> 'a -> Base.Unit.t

add_bin_prot b writer x appends the bin-protted representation of x at the end of the buffer b.

val add_substitute : t -> (Base.String.t -> Base.String.t) -> Base.String.t -> Base.Unit.t

add_substitute b f s appends the string pattern s at the end of the buffer b with substitution. The substitution process looks for variables into the pattern and substitutes each variable name by its value, as obtained by applying the mapping f to the variable name. Inside the string pattern, a variable name immediately follows a non-escaped $ character and is one of the following:

a non empty sequence of alphanumeric or _ characters,

an arbitrary sequence of characters enclosed by a pair of matching parentheses or curly brackets.

An escaped $ character is a $ that immediately follows a backslash character; it then stands for a plain $. Raise Caml.Not_found or Not_found_s if the closing character of a parenthesized variable cannot be found.

module Format : sig ... end

module Printf : sig ... end