Module Bigstring.Unstable

type nonrec t = t

include Bin_prot.Binable.S with type t := t

type t

include Bin_prot.Binable.S_only_functions with type t := t

type t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> Core_kernel__.Import.int

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

type nonrec t_frozen = t_frozen

val bin_shape_t_frozen : Bin_prot.Shape.t

val bin_size_t_frozen : t_frozen Bin_prot.Size.sizer

val bin_write_t_frozen : t_frozen Bin_prot.Write.writer

val bin_writer_t_frozen : t_frozen Bin_prot.Type_class.writer

val bin_read_t_frozen : t_frozen Bin_prot.Read.reader

val __bin_read_t_frozen__ : (Core_kernel__.Import.int -> t_frozen) Bin_prot.Read.reader

val bin_reader_t_frozen : t_frozen Bin_prot.Type_class.reader

val bin_t_frozen : t_frozen Bin_prot.Type_class.t

val compare_t_frozen : t_frozen -> t_frozen -> Core_kernel__.Import.int

val hash_fold_t_frozen : Base.Hash.state -> t_frozen -> Base.Hash.state

val hash_t_frozen : t_frozen -> Base.Hash.hash_value

val sexp_of_t_frozen : t_frozen -> Ppx_sexp_conv_lib.Sexp.t