Module type Container_intf.S1_permissions

type ('a, -'permissions) t

val mem : ('a, [> Perms.Export.read ]) t -> 'a -> equal:('a -> 'a -> Base.Bool.t) -> Base.Bool.t

Checks whether the provided element is there, using polymorphic compare if equal is not provided.

val length : (_, [> Perms.Export.read ]) t -> Base.Int.t

val is_empty : (_, [> Perms.Export.read ]) t -> Base.Bool.t

val iter : ('a, [> Perms.Export.read ]) t -> f:('a -> Base.Unit.t) -> Base.Unit.t

val fold : ('a, [> Perms.Export.read ]) t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : ('a, [> Perms.Export.read ]) t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Result.t) -> ('accum, 'e) Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : ('a, [> Perms.Export.read ]) t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

val exists : ('a, [> Perms.Export.read ]) t -> f:('a -> Base.Bool.t) -> Base.Bool.t

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : ('a, [> Perms.Export.read ]) t -> f:('a -> Base.Bool.t) -> Base.Bool.t

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : ('a, [> Perms.Export.read ]) t -> f:('a -> Base.Bool.t) -> Base.Int.t

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base.Container.Summable with type t = 'sum) -> ('a, [> Perms.Export.read ]) t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for i in the container

val find : ('a, [> Perms.Export.read ]) t -> f:('a -> Base.Bool.t) -> 'a Base.Option.t

Returns as an option the first element for which f evaluates to true.

val find_map : ('a, [> Perms.Export.read ]) t -> f:('a -> 'b Base.Option.t) -> 'b Base.Option.t

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : ('a, [> Perms.Export.read ]) t -> 'a Base.List.t

val to_array : ('a, [> Perms.Export.read ]) t -> 'a Base.Array.t

val min_elt : ('a, [> Perms.Export.read ]) t -> compare:('a -> 'a -> Base.Int.t) -> 'a Base.Option.t

Returns a min (resp max) element from the collection using the provided compare function. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold. Returns None iff the collection is empty.

val max_elt : ('a, [> Perms.Export.read ]) t -> compare:('a -> 'a -> Base.Int.t) -> 'a Base.Option.t