Module Core_kernel_private.Stable_internal

include Bin_prot.Std

include Bin_prot.Size

type 'a sizer = 'a -> int

type ('a, 'b) sizer1 = 'a sizer -> 'b sizer

type ('a, 'b, 'c) sizer2 = 'a sizer -> ('b, 'c) sizer1

type ('a, 'b, 'c, 'd) sizer3 = 'a sizer -> ('b, 'c, 'd) sizer2

val bin_size_unit : unit sizer

val bin_size_bool : bool sizer

val bin_size_string : string sizer

val bin_size_bytes : bytes sizer

val bin_size_char : char sizer

val bin_size_int : int sizer

val bin_size_float : float sizer

val bin_size_int32 : int32 sizer

val bin_size_int64 : int64 sizer

val bin_size_nativeint : nativeint sizer

val bin_size_nat0 : Bin_prot.Nat0.t sizer

val bin_size_ref : ('a, 'a Stdlib.ref) sizer1

val bin_size_lazy_t : ('a, 'a lazy_t) sizer1

val bin_size_lazy : ('a, 'a lazy_t) sizer1

val bin_size_option : ('a, 'a option) sizer1

val bin_size_pair : ('a, 'b, 'a * 'b) sizer2

val bin_size_triple : ('a, 'b, 'c, 'a * 'b * 'c) sizer3

val bin_size_list : ('a, 'a list) sizer1

val bin_size_array : ('a, 'a array) sizer1

val bin_size_hashtbl : ('a, 'b, ('a, 'b) Stdlib.Hashtbl.t) sizer2

val bin_size_float32_vec : Bin_prot.Common.vec32 sizer

val bin_size_float64_vec : Bin_prot.Common.vec64 sizer

val bin_size_vec : Bin_prot.Common.vec sizer

val bin_size_float32_mat : Bin_prot.Common.mat32 sizer

val bin_size_float64_mat : Bin_prot.Common.mat64 sizer

val bin_size_mat : Bin_prot.Common.mat sizer

val bin_size_bigstring : Bin_prot.Common.buf sizer

val bin_size_float_array : float array sizer

val bin_size_variant_int : int sizer

val bin_size_int_8bit : int sizer

val bin_size_int_16bit : int sizer

val bin_size_int_32bit : int sizer

val bin_size_int_64bit : int sizer

val bin_size_int64_bits : int64 sizer

val bin_size_network16_int : int sizer

val bin_size_network32_int : int sizer

val bin_size_network32_int32 : int32 sizer

val bin_size_network64_int : int sizer

val bin_size_network64_int64 : int64 sizer

val bin_size_array_no_length : ('a, 'a array) sizer1

val bin_size_md5 : Md5_lib.t sizer

module Maximum : sig ... end

module Minimum : sig ... end

val bin_unit : unit Bin_prot.Type_class.t

val bin_shape_unit : Bin_prot.Shape.t

val bin_writer_unit : unit Bin_prot.Type_class.writer

val bin_write_unit : unit Bin_prot.Write.writer

val bin_reader_unit : unit Bin_prot.Type_class.reader

val bin_read_unit : unit Bin_prot.Read.reader

val __bin_read_unit__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_bool : bool Bin_prot.Type_class.t

val bin_shape_bool : Bin_prot.Shape.t

val bin_writer_bool : bool Bin_prot.Type_class.writer

val bin_write_bool : bool Bin_prot.Write.writer

val bin_reader_bool : bool Bin_prot.Type_class.reader

val bin_read_bool : bool Bin_prot.Read.reader

val __bin_read_bool__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_string : string Bin_prot.Type_class.t

val bin_shape_string : Bin_prot.Shape.t

val bin_writer_string : string Bin_prot.Type_class.writer

val bin_write_string : string Bin_prot.Write.writer

val bin_reader_string : string Bin_prot.Type_class.reader

val bin_read_string : string Bin_prot.Read.reader

val __bin_read_string__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_bytes : bytes Bin_prot.Type_class.t

val bin_shape_bytes : Bin_prot.Shape.t

val bin_writer_bytes : bytes Bin_prot.Type_class.writer

val bin_write_bytes : bytes Bin_prot.Write.writer

val bin_reader_bytes : bytes Bin_prot.Type_class.reader

val bin_read_bytes : bytes Bin_prot.Read.reader

val __bin_read_bytes__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_char : char Bin_prot.Type_class.t

val bin_shape_char : Bin_prot.Shape.t

val bin_writer_char : char Bin_prot.Type_class.writer

val bin_write_char : char Bin_prot.Write.writer

val bin_reader_char : char Bin_prot.Type_class.reader

val bin_read_char : char Bin_prot.Read.reader

val __bin_read_char__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_int : int Bin_prot.Type_class.t

val bin_shape_int : Bin_prot.Shape.t

val bin_writer_int : int Bin_prot.Type_class.writer

val bin_write_int : int Bin_prot.Write.writer

val bin_reader_int : int Bin_prot.Type_class.reader

val bin_read_int : int Bin_prot.Read.reader

val __bin_read_int__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_float : float Bin_prot.Type_class.t

val bin_shape_float : Bin_prot.Shape.t

val bin_writer_float : float Bin_prot.Type_class.writer

val bin_write_float : float Bin_prot.Write.writer

val bin_reader_float : float Bin_prot.Type_class.reader

val bin_read_float : float Bin_prot.Read.reader

val __bin_read_float__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

type float_array = float array

val bin_float_array : Bin_prot.Type_class.float_array Bin_prot.Type_class.t

val bin_shape_float_array : Bin_prot.Shape.t

val bin_writer_float_array : Bin_prot.Type_class.float_array Bin_prot.Type_class.writer

val bin_write_float_array : float array Bin_prot.Write.writer

val bin_reader_float_array : Bin_prot.Type_class.float_array Bin_prot.Type_class.reader

val bin_read_float_array : float array Bin_prot.Read.reader

val __bin_read_float_array__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_int32 : int32 Bin_prot.Type_class.t

val bin_shape_int32 : Bin_prot.Shape.t

val bin_writer_int32 : int32 Bin_prot.Type_class.writer

val bin_write_int32 : int32 Bin_prot.Write.writer

val bin_reader_int32 : int32 Bin_prot.Type_class.reader

val bin_read_int32 : int32 Bin_prot.Read.reader

val __bin_read_int32__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_int64 : int64 Bin_prot.Type_class.t

val bin_shape_int64 : Bin_prot.Shape.t

val bin_writer_int64 : int64 Bin_prot.Type_class.writer

val bin_write_int64 : int64 Bin_prot.Write.writer

val bin_reader_int64 : int64 Bin_prot.Type_class.reader

val bin_read_int64 : int64 Bin_prot.Read.reader

val __bin_read_int64__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_nativeint : nativeint Bin_prot.Type_class.t

val bin_shape_nativeint : Bin_prot.Shape.t

val bin_writer_nativeint : nativeint Bin_prot.Type_class.writer

val bin_write_nativeint : nativeint Bin_prot.Write.writer

val bin_reader_nativeint : nativeint Bin_prot.Type_class.reader

val bin_read_nativeint : nativeint Bin_prot.Read.reader

val __bin_read_nativeint__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_ref : ('a, 'a Stdlib.ref) Bin_prot.Type_class.S1.t

val bin_shape_ref : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_ref : ('a, 'a Stdlib.ref) Bin_prot.Type_class.S1.writer

val bin_write_ref : ('a, 'a Stdlib.ref) Bin_prot.Write.writer1

val bin_reader_ref : ('a, 'a Stdlib.ref) Bin_prot.Type_class.S1.reader

val bin_read_ref : ('a, 'a Stdlib.ref) Bin_prot.Read.reader1

val __bin_read_ref__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'c -> 'd

val bin_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.t

val bin_shape_lazy_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.writer

val bin_write_lazy_t : ('a, 'a lazy_t) Bin_prot.Write.writer1

val bin_reader_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.reader

val bin_read_lazy_t : ('a, 'a lazy_t) Bin_prot.Read.reader1

val __bin_read_lazy_t__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'c -> 'd

val bin_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.t

val bin_shape_lazy : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.writer

val bin_write_lazy : ('a, 'a lazy_t) Bin_prot.Write.writer1

val bin_reader_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.reader

val bin_read_lazy : ('a, 'a lazy_t) Bin_prot.Read.reader1

val __bin_read_lazy__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'c -> 'd

val bin_option : ('a, 'a option) Bin_prot.Type_class.S1.t

val bin_shape_option : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_option : ('a, 'a option) Bin_prot.Type_class.S1.writer

val bin_write_option : ('a, 'a option) Bin_prot.Write.writer1

val bin_reader_option : ('a, 'a option) Bin_prot.Type_class.S1.reader

val bin_read_option : ('a, 'a option) Bin_prot.Read.reader1

val __bin_read_option__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'c -> 'd

val bin_list : ('a, 'a list) Bin_prot.Type_class.S1.t

val bin_shape_list : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_list : ('a, 'a list) Bin_prot.Type_class.S1.writer

val bin_write_list : ('a, 'a list) Bin_prot.Write.writer1

val bin_reader_list : ('a, 'a list) Bin_prot.Type_class.S1.reader

val bin_read_list : ('a, 'a list) Bin_prot.Read.reader1

val __bin_read_list__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'c -> 'd

val bin_array : ('a, 'a array) Bin_prot.Type_class.S1.t

val bin_shape_array : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_array : ('a, 'a array) Bin_prot.Type_class.S1.writer

val bin_write_array : ('a, 'a array) Bin_prot.Write.writer1

val bin_reader_array : ('a, 'a array) Bin_prot.Type_class.S1.reader

val bin_read_array : ('a, 'a array) Bin_prot.Read.reader1

val __bin_read_array__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'c -> 'd

val bin_hashtbl : ('a, 'b, ('a, 'b) Stdlib.Hashtbl.t) Bin_prot.Type_class.S2.t

val bin_shape_hashtbl : Bin_prot.Shape.t -> Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_hashtbl : ('a, 'b, ('a, 'b) Stdlib.Hashtbl.t) Bin_prot.Type_class.S2.writer

val bin_write_hashtbl : ('a, 'b, ('a, 'b) Stdlib.Hashtbl.t) Bin_prot.Write.writer2

val bin_reader_hashtbl : ('a, 'b, ('a, 'b) Stdlib.Hashtbl.t) Bin_prot.Type_class.S2.reader

val bin_read_hashtbl : ('a, 'b, ('a, 'b) Stdlib.Hashtbl.t) Bin_prot.Read.reader2

val __bin_read_hashtbl__ : 'a -> 'b -> 'c -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'd -> 'e

val bin_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.t

val bin_shape_bigstring : Bin_prot.Shape.t

val bin_writer_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.writer

val bin_write_bigstring : Bin_prot.Common.buf Bin_prot.Write.writer

val bin_reader_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.reader

val bin_read_bigstring : Bin_prot.Common.buf Bin_prot.Read.reader

val __bin_read_bigstring__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_mat : Bin_prot.Common.mat Bin_prot.Type_class.t

val bin_shape_mat : Bin_prot.Shape.t

val bin_writer_mat : Bin_prot.Common.mat Bin_prot.Type_class.writer

val bin_write_mat : Bin_prot.Common.mat Bin_prot.Write.writer

val bin_reader_mat : Bin_prot.Common.mat Bin_prot.Type_class.reader

val bin_read_mat : Bin_prot.Common.mat Bin_prot.Read.reader

val __bin_read_mat__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.t

val bin_shape_float32_mat : Bin_prot.Shape.t

val bin_writer_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.writer

val bin_write_float32_mat : Bin_prot.Common.mat32 Bin_prot.Write.writer

val bin_reader_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.reader

val bin_read_float32_mat : Bin_prot.Common.mat32 Bin_prot.Read.reader

val __bin_read_float32_mat__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.t

val bin_shape_float64_mat : Bin_prot.Shape.t

val bin_writer_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.writer

val bin_write_float64_mat : Bin_prot.Common.mat64 Bin_prot.Write.writer

val bin_reader_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.reader

val bin_read_float64_mat : Bin_prot.Common.mat64 Bin_prot.Read.reader

val __bin_read_float64_mat__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_vec : Bin_prot.Common.vec Bin_prot.Type_class.t

val bin_shape_vec : Bin_prot.Shape.t

val bin_writer_vec : Bin_prot.Common.vec Bin_prot.Type_class.writer

val bin_write_vec : Bin_prot.Common.vec Bin_prot.Write.writer

val bin_reader_vec : Bin_prot.Common.vec Bin_prot.Type_class.reader

val bin_read_vec : Bin_prot.Common.vec Bin_prot.Read.reader

val __bin_read_vec__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.t

val bin_shape_float32_vec : Bin_prot.Shape.t

val bin_writer_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.writer

val bin_write_float32_vec : Bin_prot.Common.vec32 Bin_prot.Write.writer

val bin_reader_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.reader

val bin_read_float32_vec : Bin_prot.Common.vec32 Bin_prot.Read.reader

val __bin_read_float32_vec__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

val bin_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.t

val bin_shape_float64_vec : Bin_prot.Shape.t

val bin_writer_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.writer

val bin_write_float64_vec : Bin_prot.Common.vec64 Bin_prot.Write.writer

val bin_reader_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.reader

val bin_read_float64_vec : Bin_prot.Common.vec64 Bin_prot.Read.reader

val __bin_read_float64_vec__ : 'a -> pos_ref:Bin_prot.Common.pos Stdlib.ref -> 'b -> 'c

include Base.Hash.Builtin

val hash_fold_nativeint : 'a Base.Hash.folder

val hash_fold_int64 : 'a Base.Hash.folder

val hash_fold_int32 : 'a Base.Hash.folder

val hash_fold_char : 'a Base.Hash.folder

val hash_fold_int : 'a Base.Hash.folder

val hash_fold_bool : 'a Base.Hash.folder

val hash_fold_string : 'a Base.Hash.folder

val hash_fold_float : 'a Base.Hash.folder

val hash_fold_unit : 'a Base.Hash.folder

val hash_fold_option : 'a Base.Hash.folder -> 'a Base.Hash.folder

val hash_fold_list : 'a Base.Hash.folder -> 'a Base.Hash.folder

val hash_fold_lazy_t : 'a Base.Hash.folder -> 'a Base.Hash.folder

val hash_fold_ref_frozen : 'a Base.Hash.folder -> 'a Base.Hash.folder

Hash support for array and ref is provided, but is potentially DANGEROUS, since it incorporates the current contents of the array/ref into the hash value. Because of this we add a _frozen suffix to the function name.

Hash support for string is also potentially DANGEROUS, but strings are mutated less often, so we don't append _frozen to it.

Also note that we don't support bytes.

val hash_fold_array_frozen : 'a Base.Hash.folder -> 'a Base.Hash.folder

val hash_nativeint : nativeint -> Base.Hash.hash_value

val hash_int64 : int64 -> Base.Hash.hash_value

val hash_int32 : int32 -> Base.Hash.hash_value

val hash_char : char -> Base.Hash.hash_value

val hash_int : int -> Base.Hash.hash_value

val hash_bool : bool -> Base.Hash.hash_value

val hash_string : string -> Base.Hash.hash_value

val hash_float : float -> Base.Hash.hash_value

val hash_unit : unit -> Base.Hash.hash_value

include sig ... end with type 'a Stable_internal.array := 'a Base.Array.t with type Stable_internal.bool := Base.Bool.t with type Stable_internal.char := Base.Char.t with type Stable_internal.exn := Base.Exn.t with type Stable_internal.float := Base.Float.t with type Stable_internal.int := Base.Int.t with type Stable_internal.int32 := Base.Int32.t with type Stable_internal.int64 := Base.Int64.t with type 'a Stable_internal.list := 'a Base.List.t with type Stable_internal.nativeint := Base.Nativeint.t with type 'a Stable_internal.option := 'a Base.Option.t with type 'a Stable_internal.ref := 'a Base.Ref.t with type Stable_internal.string := Base.String.t with type Stable_internal.bytes := Base.Bytes.t with type Stable_internal.unit := Base.Unit.t

val sexp_of_array : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a Base.Array.t -> Ppx_sexp_conv_lib.Sexp.t

val array_of_sexp : (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a Base.Array.t

val sexp_of_bool : Base.Bool.t -> Ppx_sexp_conv_lib.Sexp.t

val bool_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Bool.t

val sexp_of_char : Base.Char.t -> Ppx_sexp_conv_lib.Sexp.t

val char_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Char.t

val sexp_of_exn : Base.Exn.t -> Ppx_sexp_conv_lib.Sexp.t

val sexp_of_float : Base.Float.t -> Ppx_sexp_conv_lib.Sexp.t

val float_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Float.t

val sexp_of_int : Base.Int.t -> Ppx_sexp_conv_lib.Sexp.t

val int_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Int.t

val sexp_of_int32 : Base.Int32.t -> Ppx_sexp_conv_lib.Sexp.t

val int32_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Int32.t

val sexp_of_int64 : Base.Int64.t -> Ppx_sexp_conv_lib.Sexp.t

val int64_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Int64.t

val sexp_of_list : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a Base.List.t -> Ppx_sexp_conv_lib.Sexp.t

val list_of_sexp : (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a Base.List.t

val sexp_of_nativeint : Base.Nativeint.t -> Ppx_sexp_conv_lib.Sexp.t

val nativeint_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Nativeint.t

val sexp_of_option : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a Base.Option.t -> Ppx_sexp_conv_lib.Sexp.t

val option_of_sexp : (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a Base.Option.t

val sexp_of_ref : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a Base.Ref.t -> Ppx_sexp_conv_lib.Sexp.t

val ref_of_sexp : (Ppx_sexp_conv_lib.Sexp.t -> 'a) -> Ppx_sexp_conv_lib.Sexp.t -> 'a Base.Ref.t

val sexp_of_string : Base.String.t -> Ppx_sexp_conv_lib.Sexp.t

val string_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.String.t

val sexp_of_bytes : Base.Bytes.t -> Ppx_sexp_conv_lib.Sexp.t

val bytes_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Bytes.t

val sexp_of_unit : Base.Unit.t -> Ppx_sexp_conv_lib.Sexp.t

val unit_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> Base.Unit.t

include sig ... end with type 'a Stable_internal.sexp_option := 'a Base.Option.t with type 'a Stable_internal.sexp_list := 'a Base.List.t

val bin_shape_sexp_option : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_sexp_option : 'a Bin_prot.Size.sizer -> 'a Base.Option.t Bin_prot.Size.sizer

val bin_write_sexp_option : 'a Bin_prot.Write.writer -> 'a Base.Option.t Bin_prot.Write.writer

val bin_writer_sexp_option : 'a Bin_prot.Type_class.writer -> 'a Base.Option.t Bin_prot.Type_class.writer

val bin_read_sexp_option : 'a Bin_prot.Read.reader -> 'a Base.Option.t Bin_prot.Read.reader

val __bin_read_sexp_option__ : 'a Bin_prot.Read.reader -> (Base.Int.t -> 'a Base.Option.t) Bin_prot.Read.reader

val bin_reader_sexp_option : 'a Bin_prot.Type_class.reader -> 'a Base.Option.t Bin_prot.Type_class.reader

val bin_sexp_option : 'a Bin_prot.Type_class.t -> 'a Base.Option.t Bin_prot.Type_class.t

val compare_sexp_option : ('a -> 'a -> Base.Int.t) -> 'a Base.Option.t -> 'a Base.Option.t -> Base.Int.t

val hash_fold_sexp_option : (Base.Hash.state -> 'a -> Base.Hash.state) -> Base.Hash.state -> 'a Base.Option.t -> Base.Hash.state

val bin_shape_sexp_list : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_sexp_list : 'a Bin_prot.Size.sizer -> 'a Base.List.t Bin_prot.Size.sizer

val bin_write_sexp_list : 'a Bin_prot.Write.writer -> 'a Base.List.t Bin_prot.Write.writer

val bin_writer_sexp_list : 'a Bin_prot.Type_class.writer -> 'a Base.List.t Bin_prot.Type_class.writer

val bin_read_sexp_list : 'a Bin_prot.Read.reader -> 'a Base.List.t Bin_prot.Read.reader

val __bin_read_sexp_list__ : 'a Bin_prot.Read.reader -> (Base.Int.t -> 'a Base.List.t) Bin_prot.Read.reader

val bin_reader_sexp_list : 'a Bin_prot.Type_class.reader -> 'a Base.List.t Bin_prot.Type_class.reader

val bin_sexp_list : 'a Bin_prot.Type_class.t -> 'a Base.List.t Bin_prot.Type_class.t

val compare_sexp_list : ('a -> 'a -> Base.Int.t) -> 'a Base.List.t -> 'a Base.List.t -> Base.Int.t

val hash_fold_sexp_list : (Base.Hash.state -> 'a -> Base.Hash.state) -> Base.Hash.state -> 'a Base.List.t -> Base.Hash.state

type 'a sexp_option = 'a Base.Option.t

type 'a sexp_list = 'a Base.List.t