Module Utils.Of_minimal

Parameters

module S : Bin_prot.Binable.Minimal.S

Signature

include Bin_prot.Binable.S_only_functions with type t := S.t

val bin_size_t : S.t Bin_prot.Size.sizer

val bin_write_t : S.t Bin_prot.Write.writer

val bin_read_t : S.t Bin_prot.Read.reader

val __bin_read_t__ : (int -> S.t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : S.t Bin_prot.Type_class.writer

val bin_reader_t : S.t Bin_prot.Type_class.reader

val bin_t : S.t Bin_prot.Type_class.t