Module Core_kernel.Doubly_linked

module type S = sig ... end

include S

module Elt : sig ... end

type 'a t

val compare : ('a -> 'a -> Base.Int.t) -> 'a t -> 'a t -> Base.Int.t

include Ppx_sexp_conv_lib.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

include Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__Container_intf.Export.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

include Base.Invariant.S1 with type 'a t := 'a t

val invariant : ('a -> unit) -> 'a t -> unit

Creating doubly-linked lists

val create : Base.Unit.t -> 'a t

val of_list : 'a Base.List.t -> 'a t

of_list l returns a doubly-linked list t with the same elements as l and in the same order (i.e., the first element of l is the first element of t). It is always the case that l = to_list (of_list l).

val of_array : 'a Base.Array.t -> 'a t

Predicates

val equal : 'a t -> 'a t -> Base.Bool.t

pointer equality

val is_first : 'a t -> 'a Elt.t -> Base.Bool.t

val is_last : 'a t -> 'a Elt.t -> Base.Bool.t

val mem_elt : 'a t -> 'a Elt.t -> Base.Bool.t

Constant-time extraction of first and last elements

val first_elt : 'a t -> 'a Elt.t Base.Option.t

val last_elt : 'a t -> 'a Elt.t Base.Option.t

val first : 'a t -> 'a Base.Option.t

val last : 'a t -> 'a Base.Option.t

Constant-time retrieval of next or previous element

val next : 'a t -> 'a Elt.t -> 'a Elt.t Base.Option.t

val prev : 'a t -> 'a Elt.t -> 'a Elt.t Base.Option.t

Constant-time insertion of a new element

val insert_before : 'a t -> 'a Elt.t -> 'a -> 'a Elt.t

val insert_after : 'a t -> 'a Elt.t -> 'a -> 'a Elt.t

val insert_first : 'a t -> 'a -> 'a Elt.t

val insert_last : 'a t -> 'a -> 'a Elt.t

Constant-time move of an element from and to positions in the same list

val move_to_front : 'a t -> 'a Elt.t -> Base.Unit.t

val move_to_back : 'a t -> 'a Elt.t -> Base.Unit.t

val move_after : 'a t -> 'a Elt.t -> anchor:'a Elt.t -> Base.Unit.t

val move_before : 'a t -> 'a Elt.t -> anchor:'a Elt.t -> Base.Unit.t

Constant-time removal of an element

val remove : 'a t -> 'a Elt.t -> Base.Unit.t

val remove_first : 'a t -> 'a Base.Option.t

val remove_last : 'a t -> 'a Base.Option.t

val iteri : 'a t -> f:(Base.Int.t -> 'a -> Base.Unit.t) -> Base.Unit.t

val foldi : 'a t -> init:'b -> f:(Base.Int.t -> 'b -> 'a -> 'b) -> 'b

val fold_elt : 'a t -> init:'b -> f:('b -> 'a Elt.t -> 'b) -> 'b

fold_elt t ~init ~f is the same as fold, except f is called with the 'a Elt.t's from the list instead of the contained 'a values.

Note that like other iteration functions, it is an error to mutate t inside the fold. If you'd like to call remove on any of the 'a Elt.t's, use filter_inplace.

val foldi_elt : 'a t -> init:'b -> f:(Base.Int.t -> 'b -> 'a Elt.t -> 'b) -> 'b

val iter_elt : 'a t -> f:('a Elt.t -> Base.Unit.t) -> Base.Unit.t

val iteri_elt : 'a t -> f:(Base.Int.t -> 'a Elt.t -> Base.Unit.t) -> Base.Unit.t

val fold_right : 'a t -> init:'b -> f:('a -> 'b -> 'b) -> 'b

val fold_right_elt : 'a t -> init:'b -> f:('a Elt.t -> 'b -> 'b) -> 'b

val find_elt : 'a t -> f:('a -> Base.Bool.t) -> 'a Elt.t Base.Option.t

find_elt t ~f finds the first element in t that satisfies f, by testing each of element of t in turn until f succeeds.

val findi_elt : 'a t -> f:(Base.Int.t -> 'a -> Base.Bool.t) -> (Base.Int.t * 'a Elt.t) Base.Option.t

val clear : 'a t -> Base.Unit.t

clear t removes all elements from the list in constant time.

val copy : 'a t -> 'a t

val transfer : src:'a t -> dst:'a t -> Base.Unit.t

transfer ~src ~dst has the same behavior as iter src ~f:(insert_last dst); clear src except that it runs in constant time.

If s = to_list src and d = to_list dst, then after transfer ~src ~dst:

to_list src = []

to_list dst = d @ s

Linear-time mapping of lists (creates a new list)

val map : 'a t -> f:('a -> 'b) -> 'b t

val mapi : 'a t -> f:(Base.Int.t -> 'a -> 'b) -> 'b t

val filter : 'a t -> f:('a -> Base.Bool.t) -> 'a t

val filteri : 'a t -> f:(Base.Int.t -> 'a -> Base.Bool.t) -> 'a t

val filter_map : 'a t -> f:('a -> 'b Base.Option.t) -> 'b t

val filter_mapi : 'a t -> f:(Base.Int.t -> 'a -> 'b Base.Option.t) -> 'b t

Linear-time partition of lists (creates two new lists)

val partition_tf : 'a t -> f:('a -> Base.Bool.t) -> 'a t * 'a t

val partitioni_tf : 'a t -> f:(Base.Int.t -> 'a -> Base.Bool.t) -> 'a t * 'a t

val partition_map : 'a t -> f:('a -> ('b, 'c) Either.t) -> 'b t * 'c t

val partition_mapi : 'a t -> f:(Base.Int.t -> 'a -> ('b, 'c) Either.t) -> 'b t * 'c t

Linear-time in-place mapping of lists

val map_inplace : 'a t -> f:('a -> 'a) -> Base.Unit.t

map_inplace t ~f replaces all values v with f v

val mapi_inplace : 'a t -> f:(Base.Int.t -> 'a -> 'a) -> Base.Unit.t

val filter_inplace : 'a t -> f:('a -> Base.Bool.t) -> Base.Unit.t

filter_inplace t ~f removes all elements of t that don't satisfy f.

val filteri_inplace : 'a t -> f:(Base.Int.t -> 'a -> Base.Bool.t) -> Base.Unit.t

val filter_map_inplace : 'a t -> f:('a -> 'a Base.Option.t) -> Base.Unit.t

If f returns None, the element is removed, else the value is replaced with the contents of the Some

val filter_mapi_inplace : 'a t -> f:(Base.Int.t -> 'a -> 'a Base.Option.t) -> Base.Unit.t

val unchecked_iter : 'a t -> f:('a -> Base.Unit.t) -> Base.Unit.t

unchecked_iter t ~f behaves like iter t ~f except that f is allowed to modify t. Adding or removing elements before the element currently being visited has no effect on the traversal. Elements added after the element currently being visited will be traversed. Elements deleted after the element currently being visited will not be traversed. Deleting the element currently being visited is an error that is not detected (presumably leading to an infinite loop).

val to_sequence : 'a t -> 'a Sequence.t

A sequence of values from the doubly-linked list. It makes an intermediate copy of the list so that the returned sequence is immune to any subsequent mutation of the original list.