Module Identifiable.Make_plain

Parameters

module M : sig ... end

Signature

include S_common with type t := M.t

val compare : M.t -> M.t -> Base.Int.t

val hash_fold_t : Base.Hash.state -> M.t -> Base.Hash.state

val hash : M.t -> Base.Hash.hash_value

val sexp_of_t : M.t -> Ppx_sexp_conv_lib.Sexp.t

include Base.Stringable.S with type t := M.t

val of_string : string -> M.t

val to_string : M.t -> string

include Base.Pretty_printer.S with type t := M.t

val pp : Base.Formatter.t -> M.t -> unit

include Comparable.S_plain with type t := M.t

include Base.Comparable.S with type t := M.t

include Base.Comparisons.Infix with type t := M.t

val (>=) : M.t -> M.t -> bool

val (<=) : M.t -> M.t -> bool

val (=) : M.t -> M.t -> bool

val (>) : M.t -> M.t -> bool

val (<) : M.t -> M.t -> bool

val (<>) : M.t -> M.t -> bool

val equal : M.t -> M.t -> bool

val compare : M.t -> M.t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : M.t -> M.t -> M.t

val max : M.t -> M.t -> M.t

val ascending : M.t -> M.t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : M.t -> M.t -> int

val between : M.t -> low:M.t -> high:M.t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : M.t -> min:M.t -> max:M.t -> M.t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : M.t -> min:M.t -> max:M.t -> M.t Base.Or_error.t

include Base.Comparator.S with type t := M.t

type comparator_witness

val comparator : (M.t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:M.t Base.Maybe_bound.t -> M.t Base.Validate.check

val validate_ubound : max:M.t Base.Maybe_bound.t -> M.t Base.Validate.check

val validate_bound : min:M.t Base.Maybe_bound.t -> max:M.t Base.Maybe_bound.t -> M.t Base.Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : Map.S_plain with type Key.t = M.t with type Key.comparator_witness = comparator_witness

module Set : Set.S_plain with type Elt.t = M.t with type Elt.comparator_witness = comparator_witness

include Hashable.S_plain with type t := M.t

include Hashable.Common with type t := M.t

val compare : M.t -> M.t -> Base.Int.t

val hash_fold_t : Base.Hash.state -> M.t -> Base.Hash.state

val hash : M.t -> Base.Hash.hash_value

val hashable : M.t Hashtbl.Hashable.t

module Table : Hashtbl.S_plain with type key = M.t

module Hash_set : Hash_set.S_plain with type elt = M.t

module Hash_queue : Hash_queue.S with type key = M.t