Module Int63.Set

module Elt : sig ... end with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Set_intf.S with module Elt := Elt

module Tree : sig ... end

include Set_intf.S_plain with module Elt := Elt and module Tree := Tree

type t = (Elt.t, Elt.comparator_witness) Base.Set.t

val compare : t -> t -> Base.Int.t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

type named = (Elt.t, Elt.comparator_witness) Set_intf.Named.t

include Set_intf.Creators_and_accessors0 with type ('a, 'b) set := ('a, 'b) Base.Set.t with type t := t with type tree := Tree.t with type elt := Elt.t with type named := named with type comparator_witness := Elt.comparator_witness

include Set_intf.Accessors0 with type t := t with type tree := Tree.t with type elt := Elt.t with type named := named with type comparator_witness := Elt.comparator_witness

include Set_intf.Set.Accessors0 with type t := t with type tree := Tree.t with type elt := Elt.t with type named := named with type comparator_witness := Elt.comparator_witness

include Base.Container.S0 with type t := t with type elt := Elt.t

val mem : t -> Elt.t -> bool

Checks whether the provided element is there, using equality on elts.

val length : t -> int

val is_empty : t -> bool

val iter : t -> f:(Elt.t -> unit) -> unit

iter must allow exceptions raised in f to escape, terminating the iteration cleanly. The same holds for all functions below taking an f.

val fold : t -> init:'accum -> f:('accum -> Elt.t -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t.

val fold_result : t -> init:'accum -> f:('accum -> Elt.t -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : t -> init:'accum -> f:('accum -> Elt.t -> ('accum, 'final) Base__Container_intf.Export.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : t -> f:(Elt.t -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : t -> f:(Elt.t -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : t -> f:(Elt.t -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> t -> f:(Elt.t -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : t -> f:(Elt.t -> bool) -> Elt.t option

Returns as an option the first element for which f evaluates to true.

val find_map : t -> f:(Elt.t -> 'a option) -> 'a option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : t -> Elt.t list

val to_array : t -> Elt.t array

val min_elt : t -> compare:(Elt.t -> Elt.t -> int) -> Elt.t option

Returns a min (resp. max) element from the collection using the provided compare function. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold. Returns None iff the collection is empty.

val max_elt : t -> compare:(Elt.t -> Elt.t -> int) -> Elt.t option

val invariants : t -> bool

val mem : t -> Elt.t -> bool

val add : t -> Elt.t -> t

val remove : t -> Elt.t -> t

val union : t -> t -> t

val inter : t -> t -> t

val diff : t -> t -> t

val symmetric_diff : t -> t -> (Elt.t, Elt.t) Base.Either.t Base.Sequence.t

val compare_direct : t -> t -> int

val equal : t -> t -> bool

val is_subset : t -> of_:t -> bool

module Named : sig ... end

val fold_until : t -> init:'b -> f:('b -> Elt.t -> ('b, 'final) Base__Set_intf.Continue_or_stop.t) -> finish:('b -> 'final) -> 'final

val fold_right : t -> init:'b -> f:(Elt.t -> 'b -> 'b) -> 'b

val iter2 : t -> t -> f:([ `Left of Elt.t | `Right of Elt.t | `Both of Elt.t * Elt.t ] -> unit) -> unit

val filter : t -> f:(Elt.t -> bool) -> t

val partition_tf : t -> f:(Elt.t -> bool) -> t * t

val elements : t -> Elt.t list

val min_elt : t -> Elt.t option

val min_elt_exn : t -> Elt.t

val max_elt : t -> Elt.t option

val max_elt_exn : t -> Elt.t

val choose : t -> Elt.t option

val choose_exn : t -> Elt.t

val split : t -> Elt.t -> t * Elt.t option * t

val group_by : t -> equiv:(Elt.t -> Elt.t -> bool) -> t list

val find_exn : t -> f:(Elt.t -> bool) -> Elt.t

val nth : t -> int -> Elt.t option

val remove_index : t -> int -> t

val to_tree : t -> Tree.t

val to_sequence : ?⁠order:[ `Increasing | `Decreasing ] -> ?⁠greater_or_equal_to:Elt.t -> ?⁠less_or_equal_to:Elt.t -> t -> Elt.t Base.Sequence.t

val binary_search : t -> compare:(Elt.t -> 'key -> int) -> [ `Last_strictly_less_than | `Last_less_than_or_equal_to | `Last_equal_to | `First_equal_to | `First_greater_than_or_equal_to | `First_strictly_greater_than ] -> 'key -> Elt.t option

val binary_search_segmented : t -> segment_of:(Elt.t -> [ `Left | `Right ]) -> [ `Last_on_left | `First_on_right ] -> Elt.t option

val merge_to_sequence : ?⁠order:[ `Increasing | `Decreasing ] -> ?⁠greater_or_equal_to:Elt.t -> ?⁠less_or_equal_to:Elt.t -> t -> t -> (Elt.t, Elt.t) Base.Sequence.Merge_with_duplicates_element.t Base.Sequence.t

val to_map : t -> f:(Elt.t -> 'data) -> (Elt.t, 'data, Elt.comparator_witness) Base.Map.t

val quickcheck_observer : Elt.t Quickcheck.Observer.t -> t Quickcheck.Observer.t

val quickcheck_shrinker : Elt.t Quickcheck.Shrinker.t -> t Quickcheck.Shrinker.t

include Set_intf.Creators0 with type t := t with type tree := Tree.t with type elt := Elt.t with type comparator_witness := Elt.comparator_witness with type ('a, 'b) set := ('a, 'b) Base.Set.t

include Set_intf.Set.Creators0 with type t := t with type tree := Tree.t with type elt := Elt.t with type comparator_witness := Elt.comparator_witness with type ('a, 'b) set := ('a, 'b) Base.Set.t

val empty : t

val singleton : Elt.t -> t

val union_list : t list -> t

val of_list : Elt.t list -> t

val of_array : Elt.t array -> t

val of_sorted_array : Elt.t array -> t Base.Or_error.t

val of_sorted_array_unchecked : Elt.t array -> t

val of_increasing_iterator_unchecked : len:int -> f:(int -> Elt.t) -> t

val stable_dedup_list : Elt.t list -> Elt.t list

val map : ('a, 'b) Base.Set.t -> f:('a -> Elt.t) -> t

val filter_map : ('a, 'b) Base.Set.t -> f:('a -> Elt.t option) -> t

val of_tree : Tree.t -> t

val of_hash_set : Elt.t Hash_set.t -> t

val of_hashtbl_keys : (Elt.t, _) Hashtbl.t -> t

val of_map_keys : (Elt.t, _, Elt.comparator_witness) Base.Map.t -> t

val quickcheck_generator : Elt.t Quickcheck.Generator.t -> t Quickcheck.Generator.t

module Provide_of_sexp : functor (Elt : sig ... end with type Provide_of_sexp.t := Elt.t) -> sig ... end with type Provide_of_sexp.t := t

module Provide_bin_io : functor (Elt : sig ... end with type Provide_bin_io.t := Elt.t) -> Set_intf.Binable.S with type t := t

module Provide_hash : functor (Elt : Base.Hasher.S with type t := Elt.t) -> sig ... end with type Provide_hash.t := t

include Sexpable.S with type t := t

val t_of_sexp : Base.Sexp.t -> t

val sexp_of_t : t -> Base.Sexp.t

include Set_intf.Binable.S with type t := t

include Set_intf.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t