Module Core_kernel.Lazy

This module extends Base.Lazy.

type 'a t = 'a Base.Lazy.t

include Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

val compare : ('a -> 'a -> Base.Int.t) -> 'a t -> 'a t -> Base.Int.t

val hash_fold_t : (Base.Hash.state -> 'a -> Base.Hash.state) -> Base.Hash.state -> 'a t -> Base.Hash.state

include Ppx_sexp_conv_lib.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

include Typerep_lib.Typerepable.S1 with type 'a t := 'a t

val typerep_of_t : 'a Typerep_lib.Std_internal.Typerep.t -> 'a t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : 'a Typerep_lib.Typename.t -> 'a t Typerep_lib.Typename.t

include module type of Base.Lazy with type 'a t := 'a t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val hash_fold_t : (Base.Hash.state -> 'a -> Base.Hash.state) -> Base.Hash.state -> 'a t -> Base.Hash.state

include Base.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Base.Sexp.t -> 'a) -> Base.Sexp.t -> 'a t

val sexp_of_t : ('a -> Base.Sexp.t) -> 'a t -> Base.Sexp.t

include Base.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'a t) -> 'a t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ('a -> 'b) -> 'a t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'a t) -> 'a t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:('a -> 'b) -> 'a t

map t ~f is t >>| f.

val join : 'a t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> 'a t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a t

val all_unit : 'a t list -> 'a t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

exception Undefined

val force : 'a t -> 'a

force x forces the suspension x and returns its result. If x has already been forced, Lazy.force x returns the same value again without recomputing it. If it raised an exception, the same exception is raised again. Raise Undefined if the forcing of x tries to force x itself recursively.

val force_val : 'a t -> 'a

Like force except that force_val x does not use an exception handler, so it may be more efficient. However, if the computation of x raises an exception, it is unspecified whether force_val x raises the same exception or Undefined.

val from_fun : (unit -> 'a) -> 'a t

from_fun f is the same as lazy (f ()) but slightly more efficient if f is a variable. from_fun should only be used if the function f is already defined. In particular it is always less efficient to write from_fun (fun () -> expr) than lazy expr.

val from_val : 'a -> 'a t

from_val v returns an already-forced suspension of v (where v can be any expression). Essentially, from_val expr is the same as let var = expr in lazy var.

val is_val : 'a t -> bool

is_val x returns true if x has already been forced and did not raise an exception.

module T_unforcing : sig ... end

This type offers a serialization function sexp_of_t that won't force its argument. Instead, it will serialize the 'a if it is available, or just use a custom string indicating it is not forced. Note that this is not a round-trippable type, thus the type does not expose of_sexp. To be used in debug code, while tracking a Heisenbug, etc.

module Stable : sig ... end