Module Core_kernel.Month

Provides a variant type for representing months (e.g., Jan, Feb, or Nov) and functions for converting them to other formats (like an int).

type t =

| Jan

| Feb

| Mar

| Apr

| May

| Jun

| Jul

| Aug

| Sep

| Oct

| Nov

| Dec

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val jan : t

val feb : t

val mar : t

val apr : t

val may : t

val jun : t

val jul : t

val aug : t

val sep : t

val oct : t

val nov : t

val dec : t

module Variants : sig ... end

include Comparable.S_binable with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Replace_polymorphic_compare : sig ... end

include Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

module Map : Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Hashable.S_binable with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Hashtbl.Hashable.t

module Table : Hashtbl.S_binable with type key = t

module Hash_set : Hash_set.S_binable with type elt = t

module Hash_queue : Hash_queue.S with type key = t

of_string s accepts three-character abbreviations with three capitalizations (e.g. Jan, JAN, and jan).

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

val all : t Base.List.t

val of_int : Base.Int.t -> t Base.Option.t

of_int i returns the ith month if i is in 1, 2, ... , 12. Otherwise it returns None.

val of_int_exn : Base.Int.t -> t

val to_int : t -> Base.Int.t

to_int t returns an int in 1, 2, ... 12.

val shift : t -> Base.Int.t -> t

shift t i goes forward (or backward) the specified number of months.

module Export : sig ... end

module Stable : sig ... end