Module Core_kernel.Or_error

This module extends Base.Or_error with bin_io.

type 'a t = ('a, Error.t) Result.t

include Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

include module type of struct include Base.Or_error end with type 'a t := 'a t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool

val hash_fold_t : (Base.Hash.state -> 'a -> Base.Hash.state) -> Base.Hash.state -> 'a t -> Base.Hash.state

include Base.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Base.Sexp.t -> 'a) -> Base.Sexp.t -> 'a t

val sexp_of_t : ('a -> Base.Sexp.t) -> 'a t -> Base.Sexp.t

Applicative functions don't have quite the same semantics as Applicative.Of_Monad(Or_error) would give -- apply (Error e1) (Error e2) returns the combination of e1 and e2, whereas it would only return e1 if it were defined using bind.

include Base.Applicative.S with type 'a t := 'a t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'a t

val both : 'a t -> 'a t -> 'a t

val (<*>) : 'a t -> 'a t -> 'a t

same as apply

val (<*) : 'a t -> 'a t -> 'a t

val (*>) : 'a t -> 'a t -> 'a t

val (>>|) : 'a t -> ('a -> 'b) -> 'a t

val apply : 'a t -> 'a t -> 'a t

val map2 : 'a t -> 'a t -> f:('a -> 'b -> 'c) -> 'a t

val map3 : 'a t -> 'a t -> 'a t -> f:('a -> 'b -> 'c -> 'd) -> 'a t

val all : 'a t list -> 'a t

val all_unit : 'a t list -> 'a t

module Applicative_infix : sig ... end

include Base.Invariant.S1 with type 'a t := 'a t

val invariant : ('a -> unit) -> 'a t -> unit

include Base.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'a t) -> 'a t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ('a -> 'b) -> 'a t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'a t) -> 'a t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:('a -> 'b) -> 'a t

map t ~f is t >>| f.

val join : 'a t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> 'a t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a t

val all_unit : 'a t list -> 'a t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

val is_ok : 'a t -> bool

val is_error : 'a t -> bool

val ignore : 'a t -> 'a t

val try_with : ?⁠backtrace:bool -> (unit -> 'a) -> 'a t

try_with f catches exceptions thrown by f and returns them in the Result.t as an Error.t. try_with_join is like try_with, except that f can throw exceptions or return an Error directly, without ending up with a nested error; it is equivalent to Result.join (try_with f).

val try_with_join : ?⁠backtrace:bool -> (unit -> 'a t) -> 'a t

val ok : 'a t -> 'ok option

ok t returns None if t is an Error, and otherwise returns the contents of the Ok constructor.

val ok_exn : 'a t -> 'a

ok_exn t throws an exception if t is an Error, and otherwise returns the contents of the Ok constructor.

val of_exn : ?⁠backtrace:[ `Get | `This of string ] -> exn -> 'a t

of_exn ?backtrace exn is Error (Error.of_exn ?backtrace exn).

val of_exn_result : ?⁠backtrace:[ `Get | `This of string ] -> ('a, exn) Base.Result.t -> 'a t

of_exn_result ?backtrace (Ok a) = Ok a

of_exn_result ?backtrace (Error exn) = of_exn ?backtrace exn

val error : ?⁠strict:unit -> string -> 'a -> ('a -> Base.Sexp.t) -> 'a t

error is a wrapper around Error.create:

error ?strict message a sexp_of_a
= Error (Error.create ?strict message a sexp_of_a)

As with Error.create, sexp_of_a a is lazily computed when the info is converted to a sexp. So, if a is mutated in the time between the call to create and the sexp conversion, those mutations will be reflected in the sexp. Use ~strict:() to force sexp_of_a a to be computed immediately.

val error_s : Base.Sexp.t -> 'a t

val error_string : string -> 'a t

error_string message is Error (Error.of_string message).

val errorf : ('a, unit, string, 'a t) Stdlib.format4 -> 'a

errorf format arg1 arg2 ... is Error (sprintf format arg1 arg2 ...). Note that it calculates the string eagerly, so when performance matters you may want to use error instead.

val tag : 'a t -> tag:string -> 'a t

tag t ~tag is Result.map_error t ~f:(Error.tag ~tag). tag_arg is similar.

val tag_s : 'a t -> tag:Base.Sexp.t -> 'a t

val tag_arg : 'a t -> string -> 'b -> ('b -> Base.Sexp.t) -> 'a t

val unimplemented : string -> 'a t

For marking a given value as unimplemented. Typically combined with conditional compilation, where on some platforms the function is defined normally, and on some platforms it is defined as unimplemented. The supplied string should be the name of the function that is unimplemented.

val map : 'a t -> f:('a -> 'b) -> 'a t

val iter : 'a t -> f:('a -> unit) -> unit

val iter_error : 'a t -> f:(Base.Error.t -> unit) -> unit

val combine_errors : 'a t list -> 'a t

combine_errors ts returns Ok if every element in ts is Ok, else it returns Error with all the errors in ts. More precisely:

combine_errors [Ok a1; ...; Ok an] = Ok [a1; ...; an]

combine_errors [...; Error e1; ...; Error en; ...]
= Error (Error.of_list [e1; ...; en]) 

val combine_errors_unit : 'a t list -> 'a t

combine_errors_unit ts returns Ok if every element in ts is Ok (), else it returns Error with all the errors in ts, like combine_errors.

val filter_ok_at_least_one : 'a t list -> 'a t

filter_ok_at_least_one ts returns all values in ts that are Ok if there is at least one, otherwise it returns the same error as combine_errors ts.

val find_ok : 'a t list -> 'a t

find_ok ts returns the first value in ts that is Ok, otherwise it returns the same error as combine_errors ts.

val find_map_ok : 'a list -> f:('a -> 'a t) -> 'a t

find_map_ok l ~f returns the first value in l for which f returns Ok, otherwise it returns the same error as combine_errors (List.map l ~f).

module Expect_test_config : Expect_test_config.S with type 'a IO_flush.t = 'a t with type 'a IO_run.t = 'a t

module Stable : sig ... end