Module Core_kernel.Percent

A scale factor, not bounded between 0% and 100%, represented as a float.

type t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

of_string and t_of_sexp disallow nan, inf, etc.

val of_string : string -> t

val to_string : t -> string

Sexps are of the form 5bp or 0.05% or 0.0005x

val t_of_sexp : Base.Sexp.t -> t

val sexp_of_t : t -> Base.Sexp.t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Base.Comparable.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Replace_polymorphic_compare : sig ... end

include Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

module Map : Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Comparable.With_zero with type t := t

val validate_positive : t Base.Validate.check

val validate_non_negative : t Base.Validate.check

val validate_negative : t Base.Validate.check

val validate_non_positive : t Base.Validate.check

val is_positive : t -> bool

val is_non_negative : t -> bool

val is_negative : t -> bool

val is_non_positive : t -> bool

val sign : t -> Base__Sign0.t

Returns Neg, Zero, or Pos in a way consistent with the above functions.

include Robustly_comparable.S with type t := t

val (>=.) : t -> t -> bool

val (<=.) : t -> t -> bool

val (=.) : t -> t -> bool

val (>.) : t -> t -> bool

val (<.) : t -> t -> bool

val (<>.) : t -> t -> bool

val robustly_compare : t -> t -> int

include Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val (*) : t -> t -> t

val (+) : t -> t -> t

val (-) : t -> t -> t

val zero : t

val neg : t -> t

val abs : t -> t

val is_zero : t -> Base.Bool.t

val is_nan : t -> Base.Bool.t

val is_inf : t -> Base.Bool.t

val apply : t -> Base.Float.t -> Base.Float.t

apply t x multiplies the percent t by x, returning a float.

val scale : t -> Base.Float.t -> t

scale t x scales the percent t by x, returning a new t.

val of_mult : Base.Float.t -> t

of_mult 5. is 5x = 500% = 50_000bp

val to_mult : t -> Base.Float.t

val of_percentage : Base.Float.t -> t

of_percentage 5. is 5% = 0.05x = 500bp

val to_percentage : t -> Base.Float.t

val of_bp : Base.Float.t -> t

of_bp 5. is 5bp = 0.05% = 0.0005x

val to_bp : t -> Base.Float.t

val of_bp_int : Base.Int.t -> t

val to_bp_int : t -> Base.Int.t

rounds down

val round_significant : t -> significant_digits:Base.Int.t -> t

0.0123456% ~significant_digits:4 is 1.235bp

val round_decimal_mult : t -> decimal_digits:Base.Int.t -> t

0.0123456% ~decimal_digits:4 is 0.0001 = 1bp

val round_decimal_percentage : t -> decimal_digits:Base.Int.t -> t

0.0123456% ~decimal_digits:4 is 0.0123% = 1.23bp

val round_decimal_bp : t -> decimal_digits:Base.Int.t -> t

0.0123456% ~decimal_digits:4 is 1.2346bp

val t_of_sexp_allow_nan_and_inf : Sexp.t -> t

val of_string_allow_nan_and_inf : Base.String.t -> t

module Format : sig ... end

A Format.t tells Percent.format how to render a floating-point value as a string, like a printf conversion specification.

val format : t -> Format.t -> Base.String.t

val validate : t -> Base.Validate.t

val sign : t -> Sign.t

val sign_exn : t -> Sign.t

The sign of a Percent.t. Both -0. and 0. map to Zero. Raises on nan. All other values map to Neg or Pos.

module Stable : sig ... end