Module Percent.Option

The value nan cannot be represented as an Option.t

type t = private Base.Float.t

val compare : t -> t -> Base.Int.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

include Immediate_option_intf.S_without_immediate_plain with type t := t with type value := t

Constructors analogous to None and Some. If not (some_is_representable x) then some x may raise or return none.

val none : t

val some : t -> t

val some_is_representable : t -> Base.Bool.t

For some representations of immediate options, the encodings of none and some overlap. For these representations, some_is_representable value = false if value cannot be represented as an option. For example, Int.Option uses min_value to represent none. For other representations, some_is_representable always returns true.

val is_none : t -> Base.Bool.t

val is_some : t -> Base.Bool.t

val value : t -> default:t -> t

value (some x) ~default = x and value none ~default = default.

val value_exn : t -> t

value_exn (some x) = x. value_exn none raises. Unlike Option.value_exn, there is no ?message argument, so that calls to value_exn that do not raise also do not have to allocate.

val unchecked_value : t -> t

unchecked_value (some x) = x. unchecked_value none returns an unspecified value. unchecked_value t is intended as an optimization of value_exn t when is_some t is known to be true.

val to_option : t -> t Base.Option.t

val of_option : t Base.Option.t -> t

module Optional_syntax : Optional_syntax.S with type t := t with type value := t