Module Core_kernel.Set_intf

This module defines interfaces used in Set. See the Map docs for a description of the design.

This module defines module types {Creators,Accessors}{0,1,2,_generic,_with_comparator}. It uses check functors to ensure that each module type is an instance of the corresponding _generic one.

We must treat Creators and Accessors separately, because we sometimes need to choose different instantiations of their options. In particular, Set itself matches Creators2_with_comparator but Accessors2 (without comparator).

module Binable : sig ... end

module Set = Base.Set

module Tree = Set.Using_comparator.Tree

module Named = Set.Named

module Container = Base.Container

module type Elt_plain = Set.Elt_plain

module type Elt = sig ... end

module type Elt_binable = sig ... end

module Elt_bin_io : sig ... end

module Without_comparator = Set.Without_comparator

module With_comparator = Set.With_comparator

module With_first_class_module = Set.With_first_class_module

module Continue_or_stop = Container.Continue_or_stop

module Merge_to_sequence_element = Sequence.Merge_with_duplicates_element

module type Accessors_generic = sig ... end

module type Accessors0 = sig ... end

module type Accessors1 = sig ... end

module type Accessors2 = sig ... end

module type Accessors2_with_comparator = sig ... end

module Check_accessors : functor (T : sig ... end) -> functor (Tree : sig ... end) -> functor (Elt : sig ... end) -> functor (Named : sig ... end) -> functor (Cmp : sig ... end) -> functor (Options : sig ... end) -> functor (M : Accessors_generic with type ('a, 'b, 'c) options := ('a, 'b, 'c) Options.t with type ('a, 'b) t := ('a, 'b) T.t with type ('a, 'b) tree := ('a, 'b) Tree.t with type 'a elt := 'a Elt.t with type 'cmp cmp := 'cmp Cmp.t with type ('a, 'b) named := ('a, 'b) Named.t) -> sig ... end

Consistency checks (same as in Container).

module Check_accessors0 : functor (M : Accessors0) -> sig ... end

module Check_accessors1 : functor (M : Accessors1) -> sig ... end

module Check_accessors2 : functor (M : Accessors2) -> sig ... end

module Check_accessors2_with_comparator : functor (M : Accessors2_with_comparator) -> sig ... end

module type Creators_generic = sig ... end

module type Creators0 = sig ... end

module type Creators1 = sig ... end

module type Creators2 = sig ... end

module type Creators2_with_comparator = sig ... end

module Check_creators : functor (T : sig ... end) -> functor (Tree : sig ... end) -> functor (Elt : sig ... end) -> functor (Cmp : sig ... end) -> functor (Options : sig ... end) -> functor (M : Creators_generic with type ('a, 'b, 'c) options := ('a, 'b, 'c) Options.t with type ('a, 'b) t := ('a, 'b) T.t with type ('a, 'b) tree := ('a, 'b) Tree.t with type 'a elt := 'a Elt.t with type 'cmp cmp := 'cmp Cmp.t) -> sig ... end

module Check_creators0 : functor (M : Creators0) -> sig ... end

module Check_creators1 : functor (M : Creators1) -> sig ... end

module Check_creators2 : functor (M : Creators2) -> sig ... end

module Check_creators2_with_comparator : functor (M : Creators2_with_comparator) -> sig ... end

module type Creators_and_accessors_generic = sig ... end

module type Creators_and_accessors0 = sig ... end

module type Creators_and_accessors1 = sig ... end

module type Creators_and_accessors2 = sig ... end

module type Creators_and_accessors2_with_comparator = sig ... end

module Make_S_plain_tree : functor (Elt : Comparator.S) -> sig ... end

module type S_plain = sig ... end

module type S = sig ... end

module type S_binable = sig ... end