Parameter Check_creators.6-M

include Set.Creators_generic with type ('a, 'b, 'c) options := ('a, 'b, 'c) Options.t with type ('a, 'b) t := ('a, 'b) T.t with type ('a, 'b) tree := ('a, 'b) Tree.t with type 'a elt := 'a Elt.t with type 'cmp cmp := 'cmp Cmp.t

type ('a, 'cmp) set

val empty : ('a, 'b, 'c) Options.t

val singleton : ('a, 'b, 'c) Options.t

val union_list : ('a, 'b, 'c) Options.t

val of_list : ('a, 'b, 'c) Options.t

val of_array : ('a, 'b, 'c) Options.t

val of_sorted_array : ('a, 'b, 'c) Options.t

val of_sorted_array_unchecked : ('a, 'b, 'c) Options.t

val of_increasing_iterator_unchecked : ('a, 'b, 'c) Options.t

val stable_dedup_list : ('a, 'b, 'c) Options.t

val map : ('a, 'b, 'c) Options.t

The types of map and filter_map are subtle. The input set, ('a, _) set, reflects the fact that these functions take a set of *any* type, with any comparator, while the output set, ('b, 'cmp) t, reflects that the output set has the particular 'cmp of the creation function. The comparator can come in one of three ways, depending on which set module is used

Set.map -- comparator comes as an argument

Set.Poly.map -- comparator is polymorphic comparison

Foo.Set.map -- comparator is Foo.comparator

val filter_map : ('a, 'b, 'c) Options.t

val of_tree : ('a, 'b, 'c) Options.t

val of_hash_set : ('a, 'b, 'c) Options.t

val of_hashtbl_keys : ('a, 'b, 'c) Options.t

val of_map_keys : ('a Elt.t, _, 'cmp Cmp.t) Base.Map.t -> ('a, 'b) T.t

Never requires a comparator because it can get one from the input Map.t.

val quickcheck_generator : ('a, 'b, 'c) Options.t