Module type Set_intf.Creators_and_accessors2

include Accessors2

include Set.Accessors2

include Base.Container.S1_phantom_invariant

type ('a, 'phantom) t

val mem : ('a, _) t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : (_, _) t -> int

val is_empty : (_, _) t -> bool

val iter : ('a, _) t -> f:('a -> unit) -> unit

val fold : ('a, _) t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t.

val fold_result : ('a, _) t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : ('a, _) t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__Container_intf.Export.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : ('a, _) t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : ('a, _) t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : ('a, _) t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> ('a, _) t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : ('a, _) t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : ('a, _) t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : ('a, _) t -> 'a list

val to_array : ('a, _) t -> 'a array

val min_elt : ('a, _) t -> compare:('a -> 'a -> int) -> 'a option

Returns a min (resp max) element from the collection using the provided compare function. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold. Returns None iff the collection is empty.

val max_elt : ('a, _) t -> compare:('a -> 'a -> int) -> 'a option

type ('a, 'cmp) tree

val invariants : (_, _) t -> bool

val mem : ('a, _) t -> 'a -> bool

val add : ('a, 'cmp) t -> 'a -> ('a, 'cmp) t

val remove : ('a, 'cmp) t -> 'a -> ('a, 'cmp) t

val union : ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t

val inter : ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t

val diff : ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t

val symmetric_diff : ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'a) Base.Either.t Base.Sequence.t

val compare_direct : ('a, 'cmp) t -> ('a, 'cmp) t -> int

val equal : ('a, 'cmp) t -> ('a, 'cmp) t -> bool

val is_subset : ('a, 'cmp) t -> of_:('a, 'cmp) t -> bool

type ('a, 'cmp) named

module Named : sig ... end

val fold_until : ('a, _) t -> init:'b -> f:('b -> 'a -> ('b, 'final) Base__Set_intf.Continue_or_stop.t) -> finish:('b -> 'final) -> 'final

val fold_right : ('a, _) t -> init:'b -> f:('a -> 'b -> 'b) -> 'b

val iter2 : ('a, 'cmp) t -> ('a, 'cmp) t -> f:([ `Left of 'a | `Right of 'a | `Both of 'a * 'a ] -> unit) -> unit

val filter : ('a, 'cmp) t -> f:('a -> bool) -> ('a, 'cmp) t

val partition_tf : ('a, 'cmp) t -> f:('a -> bool) -> ('a, 'cmp) t * ('a, 'cmp) t

val elements : ('a, _) t -> 'a list

val min_elt : ('a, _) t -> 'a option

val min_elt_exn : ('a, _) t -> 'a

val max_elt : ('a, _) t -> 'a option

val max_elt_exn : ('a, _) t -> 'a

val choose : ('a, _) t -> 'a option

val choose_exn : ('a, _) t -> 'a

val split : ('a, 'cmp) t -> 'a -> ('a, 'cmp) t * 'a option * ('a, 'cmp) t

val group_by : ('a, 'cmp) t -> equiv:('a -> 'a -> bool) -> ('a, 'cmp) t list

val find_exn : ('a, _) t -> f:('a -> bool) -> 'a

val nth : ('a, _) t -> int -> 'a option

val remove_index : ('a, 'cmp) t -> int -> ('a, 'cmp) t

val to_tree : ('a, 'cmp) t -> ('a, 'cmp) tree

val to_sequence : ?⁠order:[ `Increasing | `Decreasing ] -> ?⁠greater_or_equal_to:'a -> ?⁠less_or_equal_to:'a -> ('a, 'cmp) t -> 'a Base.Sequence.t

val binary_search : ('a, 'cmp) t -> compare:('a -> 'key -> int) -> [ `Last_strictly_less_than | `Last_less_than_or_equal_to | `Last_equal_to | `First_equal_to | `First_greater_than_or_equal_to | `First_strictly_greater_than ] -> 'key -> 'a option

val binary_search_segmented : ('a, 'cmp) t -> segment_of:('a -> [ `Left | `Right ]) -> [ `Last_on_left | `First_on_right ] -> 'a option

val merge_to_sequence : ?⁠order:[ `Increasing | `Decreasing ] -> ?⁠greater_or_equal_to:'a -> ?⁠less_or_equal_to:'a -> ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'a) Base.Sequence.Merge_with_duplicates_element.t Base.Sequence.t

val to_map : ('a, 'cmp) t -> f:('a -> 'b) -> ('a, 'b, 'cmp) Base.Map.t

val quickcheck_observer : 'a Quickcheck.Observer.t -> ('a, 'cmp) t Quickcheck.Observer.t

val quickcheck_shrinker : 'a Quickcheck.Shrinker.t -> ('a, 'cmp) t Quickcheck.Shrinker.t

include Creators2 with type ('a, 'b) t := ('a, 'b) t with type ('a, 'b) tree := ('a, 'b) tree

include Set.Creators2 with type ('a, 'b) t := ('a, 'b) t with type ('a, 'b) tree := ('a, 'b) tree

type ('a, 'cmp) set

val empty : ('a, 'b) t

val singleton : 'a -> ('a, 'b) t

val union_list : ('a, 'b) t list -> ('a, 'b) t

val of_list : 'a list -> ('a, 'b) t

val of_array : 'a array -> ('a, 'b) t

val of_sorted_array : 'a array -> ('a, 'b) t Base.Or_error.t

val of_sorted_array_unchecked : 'a array -> ('a, 'b) t

val of_increasing_iterator_unchecked : len:int -> f:(int -> 'a) -> ('a, 'b) t

val stable_dedup_list : 'a list -> 'a list

val map : ('a, _) set -> f:('a -> 'b) -> ('a, 'b) t

val filter_map : ('a, _) set -> f:('a -> 'b option) -> ('a, 'b) t

val of_tree : ('a, 'b) tree -> ('a, 'b) t

val of_hash_set : 'a Hash_set.t -> ('a, 'b) t

val of_hashtbl_keys : ('a, _) Hashtbl.t -> ('a, 'b) t

val of_map_keys : ('a, _, 'cmp) Base.Map.t -> ('a, 'b) t

val quickcheck_generator : 'a Quickcheck.Generator.t -> ('a, 'b) t Quickcheck.Generator.t