Module Core_kernel.String

This module extends Base.String.

include module type of struct include Base.String end

val hash_fold_t : Base.Hash.state -> string -> Base.Hash.state

val hash : string -> Base.Hash.hash_value

include Base.Sexpable.S with type t := string

val t_of_sexp : Base.Sexp.t -> string

val sexp_of_t : string -> Base.Sexp.t

val sub : (string, string) Base.Blit.sub

val subo : (string, string) Base.Blit.subo

include Base.Container.S0 with type t := string with type elt = char

type elt = char

val mem : string -> elt -> bool

Checks whether the provided element is there, using equality on elts.

val length : string -> int

val is_empty : string -> bool

val iter : string -> f:(elt -> unit) -> unit

iter must allow exceptions raised in f to escape, terminating the iteration cleanly. The same holds for all functions below taking an f.

val fold : string -> init:'accum -> f:('accum -> elt -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t.

val fold_result : string -> init:'accum -> f:('accum -> elt -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : string -> init:'accum -> f:('accum -> elt -> ('accum, 'final) Base__Container_intf.Export.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : string -> f:(elt -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : string -> f:(elt -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : string -> f:(elt -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> string -> f:(elt -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : string -> f:(elt -> bool) -> elt option

Returns as an option the first element for which f evaluates to true.

val find_map : string -> f:(elt -> 'a option) -> 'a option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : string -> elt list

val to_array : string -> elt array

val min_elt : string -> compare:(elt -> elt -> int) -> elt option

Returns a min (resp. max) element from the collection using the provided compare function. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold. Returns None iff the collection is empty.

val max_elt : string -> compare:(elt -> elt -> int) -> elt option

include Base.Identifiable.S with type t := string

val hash_fold_t : Base.Hash.state -> string -> Base.Hash.state

val hash : string -> Base.Hash.hash_value

include Base.Sexpable.S with type t := string

val t_of_sexp : Base.Sexp.t -> string

val sexp_of_t : string -> Base.Sexp.t

include Base.Stringable.S with type t := string

val of_string : string -> string

val to_string : string -> string

include Base.Comparable.S with type t := string

include Base.Comparisons.Infix with type t := string

val (>=) : string -> string -> bool

val (<=) : string -> string -> bool

val (=) : string -> string -> bool

val (>) : string -> string -> bool

val (<) : string -> string -> bool

val (<>) : string -> string -> bool

val equal : string -> string -> bool

val compare : string -> string -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : string -> string -> string

val max : string -> string -> string

val ascending : string -> string -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : string -> string -> int

val between : string -> low:string -> high:string -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : string -> min:string -> max:string -> string

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : string -> min:string -> max:string -> string Base.Or_error.t

include Base.Comparator.S with type t := string

type comparator_witness

val comparator : (string, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:string Base.Maybe_bound.t -> string Base.Validate.check

val validate_ubound : max:string Base.Maybe_bound.t -> string Base.Validate.check

val validate_bound : min:string Base.Maybe_bound.t -> max:string Base.Maybe_bound.t -> string Base.Validate.check

include Base.Pretty_printer.S with type t := string

val pp : Base.Formatter.t -> string -> unit

val max_length : int

Maximum length of a string.

val length : string -> int

val get : string -> int -> char

val unsafe_get : string -> int -> char

unsafe_get t i is like get t i but does not perform bounds checking. The caller must ensure that it is a memory-safe operation.

val make : int -> char -> string

val copy : string -> string

Assuming you haven't passed -unsafe-string to the compiler, strings are immutable, so there'd be no motivation to make a copy.

val init : int -> f:(int -> char) -> string

val (^) : string -> string -> string

String append. Also available unqualified, but re-exported here for documentation purposes.

Note that a ^ b must copy both a and b into a newly-allocated result string, so a ^ b ^ c ^ ... ^ z is quadratic in the number of strings. String.concat does not have this problem -- it allocates the result buffer only once.

val concat : ?⁠sep:string -> string list -> string

Concatenates all strings in the list using separator sep (with a default separator "").

val escaped : string -> string

Special characters are represented by escape sequences, following the lexical conventions of OCaml.

val contains : ?⁠pos:int -> ?⁠len:int -> string -> char -> bool

val uppercase : string -> string

Operates on the whole string using the US-ASCII character set, e.g. uppercase "foo" = "FOO".

val lowercase : string -> string

val capitalize : string -> string

Operates on just the first character using the US-ASCII character set, e.g. capitalize "foo" = "Foo".

val uncapitalize : string -> string

val index : string -> char -> int option

index_exn and index_from_exn raise Caml.Not_found or Not_found_s when char cannot be found in s.

val index_exn : string -> char -> int

val index_from : string -> int -> char -> int option

val index_from_exn : string -> int -> char -> int

val rindex : string -> char -> int option

rindex_exn and rindex_from_exn raise Caml.Not_found or Not_found_s when char cannot be found in s.

val rindex_exn : string -> char -> int

val rindex_from : string -> int -> char -> int option

val rindex_from_exn : string -> int -> char -> int

module Search_pattern : sig ... end

Substring search and replace functions. They use the Knuth-Morris-Pratt algorithm (KMP) under the hood.

val substr_index : ?⁠pos:int -> string -> pattern:string -> int option

Substring search and replace convenience functions. They call Search_pattern.create and then forget the preprocessed pattern when the search is complete. pos < 0 or pos >= length t result in no match (hence substr_index returns None and substr_index_exn raises). may_overlap indicates whether to report overlapping matches, see Search_pattern.index_all.

val substr_index_exn : ?⁠pos:int -> string -> pattern:string -> int

val substr_index_all : string -> may_overlap:bool -> pattern:string -> int list

val substr_replace_first : ?⁠pos:int -> string -> pattern:string -> with_:string -> string

val substr_replace_all : string -> pattern:string -> with_:string -> string

As with Search_pattern.replace_all, the result may still contain pattern.

val is_substring : string -> substring:string -> bool

is_substring ~substring:"bar" "foo bar baz" is true.

val is_substring_at : string -> pos:int -> substring:string -> bool

is_substring_at "foo bar baz" ~pos:4 ~substring:"bar" is true.

val to_list_rev : string -> char list

Returns the reversed list of characters contained in a list.

val rev : string -> string

rev t returns t in reverse order.

val is_suffix : string -> suffix:string -> bool

is_suffix s ~suffix returns true if s ends with suffix.

val is_prefix : string -> prefix:string -> bool

is_prefix s ~prefix returns true if s starts with prefix.

val lsplit2_exn : string -> on:char -> string * string

If the string s contains the character on, then lsplit2_exn s ~on returns a pair containing s split around the first appearance of on (from the left). Raises Caml.Not_found or Not_found_s when on cannot be found in s.

val rsplit2_exn : string -> on:char -> string * string

If the string s contains the character on, then rsplit2_exn s ~on returns a pair containing s split around the first appearance of on (from the right). Raises Caml.Not_found or Not_found_s when on cannot be found in s.

val lsplit2 : string -> on:char -> (string * string) option

lsplit2 s ~on optionally returns s split into two strings around the first appearance of on from the left.

val rsplit2 : string -> on:char -> (string * string) option

rsplit2 s ~on optionally returns s split into two strings around the first appearance of on from the right.

val split : string -> on:char -> string list

split s ~on returns a list of substrings of s that are separated by on. Consecutive on characters will cause multiple empty strings in the result. Splitting the empty string returns a list of the empty string, not the empty list.

val split_on_chars : string -> on:char list -> string list

split_on_chars s ~on returns a list of all substrings of s that are separated by one of the chars from on. on are not grouped. So a grouping of on in the source string will produce multiple empty string splits in the result.

val split_lines : string -> string list

split_lines t returns the list of lines that comprise t. The lines do not include the trailing "\n" or "\r\n".

val lfindi : ?⁠pos:int -> string -> f:(int -> char -> bool) -> int option

lfindi ?pos t ~f returns the smallest i >= pos such that f i t.[i], if there is such an i. By default, pos = 0.

val rfindi : ?⁠pos:int -> string -> f:(int -> char -> bool) -> int option

rfindi ?pos t ~f returns the largest i <= pos such that f i t.[i], if there is such an i. By default pos = length t - 1.

val lstrip : ?⁠drop:(char -> bool) -> string -> string

lstrip ?drop s returns a string with consecutive chars satisfying drop (by default white space, e.g. tabs, spaces, newlines, and carriage returns) stripped from the beginning of s.

val rstrip : ?⁠drop:(char -> bool) -> string -> string

rstrip ?drop s returns a string with consecutive chars satisfying drop (by default white space, e.g. tabs, spaces, newlines, and carriage returns) stripped from the end of s.

val strip : ?⁠drop:(char -> bool) -> string -> string

strip ?drop s returns a string with consecutive chars satisfying drop (by default white space, e.g. tabs, spaces, newlines, and carriage returns) stripped from the beginning and end of s.

val map : string -> f:(char -> char) -> string

val mapi : string -> f:(int -> char -> char) -> string

Like map, but passes each character's index to f along with the char.

val foldi : string -> init:'a -> f:(int -> 'a -> char -> 'a) -> 'a

foldi works similarly to fold, but also passes the index of each character to f.

val concat_map : ?⁠sep:string -> string -> f:(char -> string) -> string

Like map, but allows the replacement of a single character with zero or two or more characters.

val filter : string -> f:(char -> bool) -> string

filter s ~f:predicate discards characters not satisfying predicate.

val tr : target:char -> replacement:char -> string -> string

tr ~target ~replacement s replaces every instance of target in s with replacement.

val tr_multi : target:string -> replacement:string -> (string -> string) Base.Staged.t

tr_multi ~target ~replacement returns a function that replaces every instance of a character in target with the corresponding character in replacement.

If replacement is shorter than target, it is lengthened by repeating its last character. Empty replacement is illegal unless target also is.

If target contains multiple copies of the same character, the last corresponding replacement character is used. Note that character ranges are not supported, so ~target:"a-z" means the literal characters 'a', '-', and 'z'.

val chop_suffix_exn : string -> suffix:string -> string

chop_suffix_exn s ~suffix returns s without the trailing suffix, raising Invalid_argument if suffix is not a suffix of s.

val chop_prefix_exn : string -> prefix:string -> string

chop_prefix_exn s ~prefix returns s without the leading prefix, raising Invalid_argument if prefix is not a prefix of s.

val chop_suffix : string -> suffix:string -> string option

val chop_prefix : string -> prefix:string -> string option

val suffix : string -> int -> string

suffix s n returns the longest suffix of s of length less than or equal to n.

val prefix : string -> int -> string

prefix s n returns the longest prefix of s of length less than or equal to n.

val drop_suffix : string -> int -> string

drop_suffix s n drops the longest suffix of s of length less than or equal to n.

val drop_prefix : string -> int -> string

drop_prefix s n drops the longest prefix of s of length less than or equal to n.

val concat_array : ?⁠sep:string -> string array -> string

concat_array sep ar like String.concat, but operates on arrays.

val hash : string -> int

Slightly faster hash function on strings.

val equal : string -> string -> bool

Fast equality function on strings, doesn't use compare_val.

val of_char : char -> string

val of_char_list : char list -> string

module Escaping : sig ... end

Operations for escaping and unescaping strings, with parameterized escape and escapeworthy characters. Escaping/unescaping using this module is more efficient than using Pcre. Benchmark code can be found in core/benchmarks/string_escaping.ml.

type t = string

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

module Caseless : sig ... end

val slice : t -> int -> int -> t

slice t start stop returns a new string including elements t.(start) through t.(stop-1), normalized Python-style with the exception that stop = 0 is treated as stop = length t.

val nget : t -> int -> char

nget s i gets the char at normalized position i in s.

val take_while : t -> f:(char -> bool) -> t

take_while s ~f returns the longest prefix of s satisfying for_all prefix ~f (See lstrip to drop such a prefix)

val rtake_while : t -> f:(char -> bool) -> t

rtake_while s ~f returns the longest suffix of s satisfying for_all suffix ~f (See rstrip to drop such a suffix)

include Hexdump.S with type t := t

module Hexdump : sig ... end

include Identifiable.S with type t := t and type comparator_witness := comparator_witness

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Identifiable.S_common with type t := t

val compare : t -> t -> Base.Int.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Base.Pretty_printer.S with type t := t

val pp : Base.Formatter.t -> t -> unit

include Comparable.S_binable with type t := t with type comparator_witness := comparator_witness

include Base.Comparable.S with type t := t with type comparator_witness := comparator_witness

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Replace_polymorphic_compare : sig ... end

include Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

module Map : Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Hashable.S_binable with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Hashtbl.Hashable.t

module Table : Hashtbl.S_binable with type key = t

module Hash_set : Hash_set.S_binable with type elt = t

module Hash_queue : Hash_queue.S with type key = t

include Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val gen_nonempty : t Quickcheck.Generator.t

Like gen, but without empty strings.

val gen' : char Quickcheck.Generator.t -> t Quickcheck.Generator.t

Like gen, but generate strings with the given distribution of characters.

val gen_nonempty' : char Quickcheck.Generator.t -> t Quickcheck.Generator.t

Like gen', but without empty strings.

val gen_with_length : int -> char Quickcheck.Generator.t -> t Quickcheck.Generator.t

Like gen', but generate strings with the given length.

module Stable : sig ... end

Note that string is already stable by itself, since as a primitive type it is an integral part of the sexp / bin_io protocol. String.Stable exists only to introduce String.Stable.Set, String.Stable.Map, String.Stable.Table, and provide interface uniformity with other stable types.