Module Time.Zone

type t

The type of a time-zone.

bin_io and sexp representations of Zone.t are the name of the zone, and not the full data that is read from disk when Zone.find is called. The full Zone.t is reconstructed on the receiving/reading side by reloading the zone file from disk. Any zone name that is accepted by find is acceptable in the bin_io and sexp representations.

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val compare : t -> t -> Base.Int.t

val input_tz_file : zonename:Base.String.t -> filename:Base.String.t -> t

input_tz_file ~zonename ~filename read in filename and return t with name t = zonename

val likely_machine_zones : Base.String.t Base.List.t Base.Ref.t

likely_machine_zones is a list of zone names that will be searched first when trying to determine the machine zone of a box. Setting this to a likely set of zones for your application will speed the very first use of the local timezone.

val of_utc_offset : hours:Base.Int.t -> t

of_utc_offset offset returns a timezone with a static UTC offset (given in hours).

val utc : t

utc the UTC time zone. Included for convenience

val name : t -> Base.String.t

val original_filename : t -> Base.String.t Base.Option.t

original_filename t return the filename t was loaded from (if any)

val digest : t -> Md5.t Base.Option.t

digest t return the MD5 digest of the file the t was created from (if any)

val reset_transition_cache : t -> Base.Unit.t

For performance testing only; reset_transition_cache t resets an internal cache in t used to speed up repeated lookups of the same clock shift transition.

module Index : sig ... end

A time zone index refers to a range of times delimited by DST transitions at one or both ends. Every time belongs to exactly one such range. The times of DST transitions themselves belong to the range for which they are the lower bound.

val index : t -> Base.Float.t -> Index.t

Gets the index of a time.

val index_of_date_and_ofday : t -> Base.Float.t -> Index.t

val index_offset_from_utc_exn : t -> Index.t -> Base.Float.t

Gets the UTC offset of times in a specific range.

This can raise if you use an Index.t that is out of bounds for this t.

val index_abbreviation_exn : t -> Index.t -> Base.String.t

index_abbreviation_exn t index returns the abbreviation name (such as EDT, EST, JST) of given zone t for the range of index. This string conversion is one-way only, and cannot reliably be turned back into a t. This function reads and writes the zone's cached index. Raises if index is out of bounds for t.

val index_has_prev_clock_shift : t -> Index.t -> Base.Bool.t

Accessors for the DST transitions delimiting the start and end of a range, if any. The _exn accessors raise if there is no such transition. These accessors are split up to increase performance and improve allocation; they are intended as a low-level back-end for commonly-used time conversion functions. See Time.Zone and Time_ns.Zone for higher-level accessors that return an optional tuple for clock shifts in either direction.

val index_prev_clock_shift_time_exn : t -> Index.t -> Base.Float.t

val index_prev_clock_shift_amount_exn : t -> Index.t -> Base.Float.t

val index_has_next_clock_shift : t -> Index.t -> Base.Bool.t

val index_next_clock_shift_time_exn : t -> Index.t -> Base.Float.t

val index_next_clock_shift_amount_exn : t -> Index.t -> Base.Float.t

val abbreviation : t -> Base.Float.t -> Base.String.t

abbreviation t time returns the abbreviation name (such as EDT, EST, JST) of given zone t at time. This string conversion is one-way only, and cannot reliably be turned back into a t. This function reads and writes the zone's cached index.

val absolute_time_of_date_and_ofday : t -> Base.Float.t -> Base.Float.t

absolute_time_of_date_and_ofday and date_and_ofday_of_absolute_time convert between absolute times and date + ofday forms. These are low level functions not intended for most clients. These functions read and write the zone's cached index.

val date_and_ofday_of_absolute_time : t -> Base.Float.t -> Base.Float.t

val next_clock_shift : t -> strictly_after:Base.Float.t -> (Base.Float.t * Base.Float.t) Base.Option.t

Takes a Time.t and returns the next Time.t strictly after it, if any, that the time zone UTC offset changes, and by how much it does so.

val prev_clock_shift : t -> at_or_before:Base.Float.t -> (Base.Float.t * Base.Float.t) Base.Option.t

As next_clock_shift, but *at or before* the given time.