Module Arg_type.Export

Values to include in other namespaces.

val string : Core_kernel__.Import.string t

val int : Core_kernel__.Import.int t

Beware that an anonymous argument of type int cannot be specified as negative, as it is ambiguous whether -1 is a negative number or a flag. (The same applies to float, time_span, etc.) You can use the special built-in "-anon" flag to force a string starting with a hyphen to be interpreted as an anonymous argument rather than as a flag, or you can just make it a parameter to a flag to avoid the issue.

val char : Core_kernel__.Import.char t

val float : Core_kernel__.Import.float t

val bool : Core_kernel__.Import.bool t

val date : Core_kernel.Date.t t

val percent : Core_kernel.Percent.t t

val host_and_port : Core_kernel.Host_and_port.t t

val sexp : Core_kernel.Sexp.t t

val sexp_conv : (Core_kernel.Sexp.t -> 'a) -> 'a t