Module Core_kernel__.Hash_set

type 'a t = 'a Base.Hash_set.t

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

include Core_kernel__.Hash_set_intf.Creators_generic with type 'a t := 'a t with type 'a elt = 'a with type ('key, 'z) create_options := ('key, 'z) Core_kernel__.Hash_set_intf.create_options

type 'a t

type 'a elt = 'a

type ('a, 'z) create_options

val create : ('a, unit -> 'a t) create_options

val of_list : ('a, 'a elt list -> 'a t) create_options

include Core_kernel__.Hash_set_intf.Accessors with type 'a t := 'a t with type 'a elt := 'a elt

include Base.Container.Generic

type 'a t

type 'a elt

val length : _ t -> int

val is_empty : _ t -> bool

val iter : 'a t -> f:('a elt -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a elt -> 'accum) -> 'accum

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a elt -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a elt -> ('accum, 'final) Base__.Container_intf.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

val exists : 'a t -> f:('a elt -> bool) -> bool

val for_all : 'a t -> f:('a elt -> bool) -> bool

val count : 'a t -> f:('a elt -> bool) -> int

val sum : (module Base__.Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a elt -> 'sum) -> 'sum

val find : 'a t -> f:('a elt -> bool) -> 'a elt option

val find_map : 'a t -> f:('a elt -> 'b option) -> 'b option

val to_list : 'a t -> 'a elt list

val to_array : 'a t -> 'a elt array

val min_elt : 'a t -> compare:('a elt -> 'a elt -> int) -> 'a elt option

val max_elt : 'a t -> compare:('a elt -> 'a elt -> int) -> 'a elt option

val mem : 'a t -> 'a -> bool

override Container.Generic.mem

val copy : 'a t -> 'a t

preserves the equality function

val add : 'a t -> 'a -> unit

val strict_add : 'a t -> 'a -> unit Base.Or_error.t

strict_add t x returns Ok () if the x was not in t, or an Error if it was.

val strict_add_exn : 'a t -> 'a -> unit

val remove : 'a t -> 'a -> unit

val strict_remove : 'a t -> 'a -> unit Base.Or_error.t

strict_remove t x returns Ok () if the x was in t, or an Error if it was not.

val strict_remove_exn : 'a t -> 'a -> unit

val clear : 'a t -> unit

val equal : 'a t -> 'a t -> bool

val filter : 'a t -> f:('a -> bool) -> 'a t

val filter_inplace : 'a t -> f:('a -> bool) -> unit

val inter : 'key t -> 'key t -> 'key t

inter t1 t2 computes the set intersection of t1 and t2. Runs in O(min(length t1, length t2)). Behavior is undefined if t1 and t2 don't have the same equality function.

val diff : 'a t -> 'a t -> 'a t

val of_hashtbl_keys : ('a, _) Base.Hashtbl.t -> 'a t

val to_hashtbl : 'key t -> f:('key -> 'data) -> ('key, 'data) Base.Hashtbl.t

val hashable : 'key t -> 'key Core_kernel.Hashtbl.Hashable.t

module type Elt_plain = Core_kernel.Hashtbl.Key_plain

module type Elt = Core_kernel.Hashtbl.Key

module type Elt_binable = Core_kernel.Hashtbl.Key_binable

module type S_plain = Core_kernel__.Hash_set_intf.S_plain with type 'a hash_set := 'a t

module type S = Core_kernel__.Hash_set_intf.S with type 'a hash_set := 'a t

module type S_binable = Core_kernel__.Hash_set_intf.S_binable with type 'a hash_set := 'a t

module Using_hashable : sig ... end

module Poly : sig ... end

A hash set that uses polymorphic comparison.

module Make_plain : functor (Elt : Elt_plain) -> S_plain with type elt = Elt.t

module Make : functor (Elt : Elt) -> S with type elt = Elt.t

module Make_binable : functor (Elt : Elt_binable) -> S_binable with type elt = Elt.t

module M : functor (Elt : Core_kernel__.T.T) -> sig ... end

module type Sexp_of_m = sig ... end

module type M_of_sexp = sig ... end

val sexp_of_m__t : (module Sexp_of_m with type t = 'elt) -> 'elt t -> Base.Sexp.t

val m__t_of_sexp : (module M_of_sexp with type t = 'elt) -> Base.Sexp.t -> 'elt t