Module Core_kernel__.Result

type ('a, 'b) t = ('a, 'b) Base.Result.t =

| Ok of 'a

| Error of 'b

include Bin_prot.Binable.S2 with type ('a, 'b) t := ('a, 'b) t

type ('a, 'b) t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'b, ('a, 'b) t) Bin_prot.Size.sizer2

val bin_write_t : ('a, 'b, ('a, 'b) t) Bin_prot.Write.writer2

val bin_read_t : ('a, 'b, ('a, 'b) t) Bin_prot.Read.reader2

val __bin_read_t__ : ('a, 'b, int -> ('a, 'b) t) Bin_prot.Read.reader2

val bin_writer_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.writer

val bin_reader_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.reader

val bin_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.t

val compare : ('a -> 'a -> Core_kernel__.Import.int) -> ('b -> 'b -> Core_kernel__.Import.int) -> ('a, 'b) t -> ('a, 'b) t -> Core_kernel__.Import.int

val hash_fold_t : (Base.Hash.state -> 'a -> Base.Hash.state) -> (Base.Hash.state -> 'b -> Base.Hash.state) -> Base.Hash.state -> ('a, 'b) t -> Base.Hash.state

include Ppx_sexp_conv_lib.Sexpable.S2 with type ('a, 'b) t := ('a, 'b) t

type ('a, 'b) t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> ('a, 'b) t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> ('a, 'b) t -> Sexplib0.Sexp.t

include module type of Base.Result with type ('a, 'b) Result.t := ('a, 'b) t

Results

type ('a, 'e) t = ('a, 'e) Stdlib.result =

| Ok of 'a

| Error of 'e

The type for result values. Either a value Ok v or an error Error e.

val ok : 'a -> ('a, 'e) Stdlib.result

ok v is Ok v.

val error : 'e -> ('a, 'e) Stdlib.result

error e is Error e.

val value : ('a, 'e) Stdlib.result -> default:'a -> 'a

value r ~default is v if r is Ok v and default otherwise.

val get_ok : ('a, 'e) Stdlib.result -> 'a

get_ok r is v if r is Ok v and

raises Invalid_argument

otherwise.

val get_error : ('a, 'e) Stdlib.result -> 'e

get_error r is e if r is Error e and

raises Invalid_argument

otherwise.

val bind : ('a, 'e) Stdlib.result -> ('a -> ('b, 'e) Stdlib.result) -> ('b, 'e) Stdlib.result

bind r f is f v if r is Ok v and r if r is Error _.

val join : (('a, 'e) Stdlib.result, 'e) Stdlib.result -> ('a, 'e) Stdlib.result

join rr is r if rr is Ok r and rr if rr is Error _.

val map : ('a -> 'b) -> ('a, 'e) Stdlib.result -> ('b, 'e) Stdlib.result

map f r is Ok (f v) if r is Ok v and r if r is Error _.

val map_error : ('e -> 'f) -> ('a, 'e) Stdlib.result -> ('a, 'f) Stdlib.result

map_error f r is Error (f e) if r is Error e and r if r is Ok _.

val fold : ok:('a -> 'c) -> error:('e -> 'c) -> ('a, 'e) Stdlib.result -> 'c

fold ~ok ~error r is ok v if r is Ok v and error e if r is Error e.

val iter : ('a -> unit) -> ('a, 'e) Stdlib.result -> unit

iter f r is f v if r is Ok v and () otherwise.

val iter_error : ('e -> unit) -> ('a, 'e) Stdlib.result -> unit

iter_error f r is f e if r is Error e and () otherwise.

Predicates and comparisons

val is_ok : ('a, 'e) Stdlib.result -> bool

is_ok r is true iff r is Ok _.

val is_error : ('a, 'e) Stdlib.result -> bool

is_error r is true iff r is Error _.

val equal : ok:('a -> 'a -> bool) -> error:('e -> 'e -> bool) -> ('a, 'e) Stdlib.result -> ('a, 'e) Stdlib.result -> bool

equal ~ok ~error r0 r1 tests equality of r0 and r1 using ok and error to respectively compare values wrapped by Ok _ and Error _.

val compare : ok:('a -> 'a -> int) -> error:('e -> 'e -> int) -> ('a, 'e) Stdlib.result -> ('a, 'e) Stdlib.result -> int

compare ~ok ~error r0 r1 totally orders r0 and r1 using ok and error to respectively compare values wrapped by Ok _ and Error _. Ok _ values are smaller than Error _ values.

Converting

val to_option : ('a, 'e) Stdlib.result -> 'a option

to_option r is r as an option, mapping Ok v to Some v and Error _ to None.

val to_list : ('a, 'e) Stdlib.result -> 'a list

to_list r is [v] if r is Ok v and [] otherwise.

val to_seq : ('a, 'e) Stdlib.result -> 'a Stdlib.Seq.t

to_seq r is r as a sequence. Ok v is the singleton sequence containing v and Error _ is the empty sequence.

module Stable : sig ... end