Module Hexdump.Pretty

[%sexp_of: Hexdump.Pretty.t] is the same as [%sexp_of: Hexdump.t], unless the underlying sequence of characters is entirely printable. In that case, it is rendered directly as a string. This allows e.g. test output to be much more compact in the common (printable) case while still being interpretable for any byte sequence.

type nonrec t = t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t