Module type Time_intf.Basic

module Time : Core_kernel__.Time0_intf.S

include module type of sig ... end

type underlying = Time.underlying

type t = Time.t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

module Span = Time.Span

module Ofday = Time.Ofday

module Replace_polymorphic_compare = Time.Replace_polymorphic_compare

val (>=.) : t -> t -> bool

val (<=.) : t -> t -> bool

val (=.) : t -> t -> bool

val (>.) : t -> t -> bool

val (<.) : t -> t -> bool

val (<>.) : t -> t -> bool

val robustly_compare : t -> t -> int

val add : t -> Span.t -> t

val sub : t -> Span.t -> t

val diff : t -> t -> Span.t

val next : t -> t

val prev : t -> t

val to_span_since_epoch : t -> Span.t

val of_span_since_epoch : Span.t -> t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

val clamp_exn : t -> min:t -> max:t -> t

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

type comparator_witness = Time.comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Date_and_ofday = Time.Date_and_ofday

val next_multiple : ?⁠can_equal_after:Core_kernel__.Import.bool -> base:t -> after:t -> interval:Span.t -> Core_kernel__.Import.unit -> t

val prev_multiple : ?⁠can_equal_before:Core_kernel__.Import.bool -> base:t -> before:t -> interval:Span.t -> Core_kernel__.Import.unit -> t

val now : Core_kernel__.Import.unit -> t

val now : Core_kernel__.Import.unit -> t

now () returns a t representing the current time

module Zone : Zone with module Time := Time

Basic operations on times

val add : t -> Span.t -> t

add t s adds the span s to time t and returns the resulting time.

NOTE: adding spans as a means of adding days is not accurate, and may run into trouble due to shifts in daylight savings time, float arithmetic issues, and leap seconds. See the comment at the top of Zone.mli for a more complete discussion of some of the issues of time-keeping. For spans that cross date boundaries, use date functions instead.

val sub : t -> Span.t -> t

sub t s subtracts the span s from time t and returns the resulting time. See important note for add.

val diff : t -> t -> Span.t

diff t1 t2 returns time t1 minus time t2.

val abs_diff : t -> t -> Span.t

abs_diff t1 t2 returns the absolute span of time t1 minus time t2.

Comparisons

val is_earlier : t -> than:t -> Core_kernel__.Import.bool

val is_later : t -> than:t -> Core_kernel__.Import.bool