Module Shape.Base_info

type grammar =

| Zero

| One of Core_kernel__.Import.string

| Many of grammar

| Maybe of grammar

| Concat of grammar Core_kernel__.Import.list

| Ad_hoc of Core_kernel__.Import.string

val bin_shape_grammar : Bin_prot.Shape.t

val bin_size_grammar : grammar Bin_prot.Size.sizer

val bin_write_grammar : grammar Bin_prot.Write.writer

val bin_writer_grammar : grammar Bin_prot.Type_class.writer

val bin_read_grammar : grammar Bin_prot.Read.reader

val __bin_read_grammar__ : (Core_kernel__.Import.int -> grammar) Bin_prot.Read.reader

val bin_reader_grammar : grammar Bin_prot.Type_class.reader

val bin_grammar : grammar Bin_prot.Type_class.t

val compare_grammar : grammar -> grammar -> Core_kernel__.Import.int

val sexp_of_grammar : grammar -> Ppx_sexp_conv_lib.Sexp.t

val grammar_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> grammar

type anons =

| Usage of Core_kernel__.Import.string
When exec'ing an older binary whose help sexp doesn't expose the grammar.

| Grammar of grammar

val bin_shape_anons : Bin_prot.Shape.t

val bin_size_anons : anons Bin_prot.Size.sizer

val bin_write_anons : anons Bin_prot.Write.writer

val bin_writer_anons : anons Bin_prot.Type_class.writer

val bin_read_anons : anons Bin_prot.Read.reader

val __bin_read_anons__ : (Core_kernel__.Import.int -> anons) Bin_prot.Read.reader

val bin_reader_anons : anons Bin_prot.Type_class.reader

val bin_anons : anons Bin_prot.Type_class.t

val compare_anons : anons -> anons -> Core_kernel__.Import.int

val sexp_of_anons : anons -> Ppx_sexp_conv_lib.Sexp.t

val anons_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> anons

type t = {

summary : Core_kernel__.Import.string;

readme : Core_kernel__.Import.string Core_kernel__.Std_internal.sexp_option;

anons : anons;

flags : Flag_info.t Core_kernel__.Import.list;

}

include Bin_prot.Binable.S with type t := t

type t

include Bin_prot.Binable.S_only_functions with type t := t

type t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> Core_kernel__.Import.int

val flags : t -> Flag_info.t Core_kernel__.Import.list

val anons : t -> anons

val readme : t -> Core_kernel__.Import.string Core_kernel__.Std_internal.sexp_option

val summary : t -> Core_kernel__.Import.string

module Fields : sig ... end

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

type t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t