Module Core_kernel__Deque

A double-ended queue that can shrink and expand on both ends.

An index is assigned to an element when it enters the queue, and the index of an element is static (i.e., an index refers to a distinct element until that element is removed from the queue, no matter how many intervening push/pop operations occur).

One consequence of this is that the minimum index may be less than zero.

The "front" is the smallest valid index, while the "back" is the largest.

All operations are amortized O(1) with a small constant.