Module Core_kernel__Float_with_finite_only_serialization

An alias to the Float.t type that causes the sexp and bin-io serializers to fail when provided with nan or infinity.

Note that while it makes sense to use this on the definition of a type in the ml file, where it will influence the construction of the sexp and bin-io serializers, it does not make sense to use this in an mli, since it makes no guarantee at that level.