Module type Core_kernel__Info_intf.Info

include module type of sig ... end with module Internal_repr := Base.Info.Internal_repr

module type S = Base__.Info_intf.S

type t = Base__Info.t

val compare : t -> t -> int

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val t_of_sexp : Base.Sexp.t -> t

val sexp_of_t : t -> Base.Sexp.t

val invariant : t Base__.Invariant_intf.inv

val to_string_hum : t -> string

val to_string_mach : t -> string

val to_string_hum_deprecated : t -> string

val of_string : string -> t

val of_lazy : string Base.Lazy.t -> t

val of_thunk : (unit -> string) -> t

val of_lazy_t : t Base.Lazy.t -> t

val create : ?⁠here:Base__.Source_code_position0.t -> ?⁠strict:unit -> string -> 'a -> ('a -> Base.Sexp.t) -> t

val create_s : Base.Sexp.t -> t

val createf : ('a, unit, string, t) Stdlib.format4 -> 'a

val tag : t -> tag:string -> t

val tag_arg : t -> string -> 'a -> ('a -> Base.Sexp.t) -> t

val of_list : ?⁠trunc_after:int -> t list -> t

val of_exn : ?⁠backtrace:[ `Get | `This of string ] -> exn -> t

val to_exn : t -> exn

val pp : Base.Formatter.t -> t -> unit

module Internal_repr = Base__Info.Internal_repr

module Internal_repr : module type of Base.Info.Internal_repr

include Extension with type t := t

type t

include Bin_prot.Binable.S with type t := t

type t

include Bin_prot.Binable.S_only_functions with type t := t

type t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

module Stable : sig ... end

module Extend : functor (Info : Base.Info.S) -> Extension with type t := Info.t