Module Core_kernel__Map

Map is a functional data structure (balanced binary tree) implementing finite maps over a totally-ordered domain, called a "key".

For example:

let empty = Map.empty (module String)
let numbers =
  Map.of_alist_exn (module String)
    ["three", Substr "three"; "four", Substr "four"]

Note that the functions in Map are polymorphic over the type of the key and of the data; you just need to pass in the first-class module for the key type (here, String).

Suppose you wanted to define a new module Foo to use in a map. You would write:

module Foo = struct
  module T = struct
    type t = int * int
    let compare x y = Tuple2.compare Int.compare Int.compare
    let sexp_of_t = Tuple2.sexp_of_t Int.sexp_of_t Int.sexp_of_t
  end
  include T
  include Comparable.Make(T)
end

This gives you a module Foo with the appropriate comparator in it, and then this:

let m = Map.empty (module Foo)

lets you create a map keyed by Foo. The reason you need to write a sexp-converter and a comparison function for this to work is that maps both need comparison and the ability to serialize the key for generating useful errors. It's yet nicer to do this with the appropriate PPXs:

module Foo = struct
  module T =
  struct  type t = int * int [@@deriving sexp_of, compare]  end
  include T
  include Comparable.Make(T)
end

The interface