Module Core_kernel__Set

This module defines the Set module for Core. Functions that construct a set take as an argument the comparator for the element type.

This module uses the same organizational approach as Map.