Module Core_kernel__Set.Merge_to_sequence_element

Produces the elements of the two sets between greater_or_equal_to and less_or_equal_to in order, noting whether each element appears in the left set, the right set, or both. In the both case, both elements are returned, in case the caller can distinguish between elements that are equal to the sets' comparator. Runs in O(length t + length t').

type ('a, 'b) t = ('a, 'b) Core_kernel.Sequence.Merge_with_duplicates_element.t =

| Left of 'a

| Right of 'b

| Both of 'a * 'b

include Bin_prot.Binable.S2 with type ('a, 'b) t := ('a, 'b) t

type ('a, 'b) t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'b, ('a, 'b) t) Bin_prot.Size.sizer2

val bin_write_t : ('a, 'b, ('a, 'b) t) Bin_prot.Write.writer2

val bin_read_t : ('a, 'b, ('a, 'b) t) Bin_prot.Read.reader2

val __bin_read_t__ : ('a, 'b, int -> ('a, 'b) t) Bin_prot.Read.reader2

val bin_writer_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.writer

val bin_reader_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.reader

val bin_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.t

val compare : ('a -> 'a -> Core_kernel__.Import.int) -> ('b -> 'b -> Core_kernel__.Import.int) -> ('a, 'b) t -> ('a, 'b) t -> Core_kernel__.Import.int

include Ppx_sexp_conv_lib.Sexpable.S2 with type ('a, 'b) t := ('a, 'b) t

type ('a, 'b) t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> ('a, 'b) t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> ('a, 'b) t -> Sexplib0.Sexp.t