Module Core_kernel__Stable_unit_test

The tests generated by these functors are run like any other unit tests: by the inline test runner when the functor is applied.

See stable_unit_test_intf.ml for documentation regarding the argument module types.