Module Core_kernel__Std_kernel

Core_kernel greatly expands the functionality available in Base while still remaining platform-agnostic. Core_kernel changes more frequently (i.e., is less stable) than Base.

Some modules are mere extensions of their counterparts in Base, usually adding generic functionality by including functors that make them binable, comparable, sexpable, blitable, etc. The bulk of Core_kernel, though, is modules providing entirely new functionality.

It is broken in two pieces, Std_kernel and Std, where the first includes modules that aren't overridden by Core, and the second defines modules that are.