Module type Core_kernel__Time_intf.Shared

type t

module Span : sig ... end

module Ofday : sig ... end

Conversions

val of_date_ofday : zone:Core_kernel__.Zone.t -> Date.t -> Ofday.t -> t

val of_date_ofday_precise : Date.t -> Ofday.t -> zone:Core_kernel__.Zone.t -> [ `Once of t | `Twice of t * t | `Never of t ]

Because timezone offsets change throughout the year (clocks go forward or back) some local times can occur twice or not at all. In the case that they occur twice, this function gives `Twice with both occurrences in order; if they do not occur at all, this function gives `Never with the time at which the local clock skips over the desired time of day.

Note that this is really only intended to work with DST transitions and not unusual or dramatic changes, like the calendar change in 1752 (run "cal 9 1752" in a shell to see). In particular it makes the assumption that midnight of each day is unambiguous.

Most callers should use of_date_ofday rather than this function. In the `Twice and `Never cases, of_date_ofday will return reasonable times for most uses.

val to_date_ofday : t -> zone:Core_kernel__.Zone.t -> Date.t * Ofday.t

val to_date_ofday_precise : t -> zone:Core_kernel__.Zone.t -> Date.t * Ofday.t * [ `Only | `Also_at of t | `Also_skipped of Date.t * Ofday.t ]

Always returns the Date.t * Ofday.t that to_date_ofday would have returned, and in addition returns a variant indicating whether the time is associated with a time zone transition.

      - `Only         -> there is a one-to-one mapping between [t]'s and
                         [Date.t * Ofday.t] pairs
      - `Also_at      -> there is another [t] that maps to the same [Date.t * Ofday.t]
                         (this date/time pair happened twice because the clock fell back)
      - `Also_skipped -> there is another [Date.t * Ofday.t] pair that never happened (due
                         to a jump forward) that [of_date_ofday] would map to the same
                         [t].

val to_date : t -> zone:Core_kernel__.Zone.t -> Date.t

val to_ofday : t -> zone:Core_kernel__.Zone.t -> Ofday.t

val reset_date_cache : Core_kernel__.Import.unit -> Core_kernel__.Import.unit

For performance testing only; reset_date_cache () resets an internal cache used to speed up to_date and related functions when called repeatedly on times that fall within the same day.

val epoch : t

midnight, Jan 1, 1970 in UTC

val convert : from_tz:Core_kernel__.Zone.t -> to_tz:Core_kernel__.Zone.t -> Date.t -> Ofday.t -> Date.t * Ofday.t

It's unspecified what happens if the given date/ofday/zone correspond to more than one date/ofday pair in the other zone.

val utc_offset : t -> zone:Core_kernel__.Zone.t -> Span.t

Other string conversions

The {to,of}_string functions in Time convert to UTC time, because a local time zone is not necessarily available. They are generous in what they will read in.

include Core_kernel__.Std_internal.Stringable with type t := t

type t

val of_string : string -> t

val to_string : t -> string

val to_filename_string : t -> zone:Core_kernel__.Zone.t -> Core_kernel__.Import.string

to_filename_string t ~zone converts t to string with format YYYY-MM-DD_HH-MM-SS.mmm which is suitable for using in filenames.

val of_filename_string : Core_kernel__.Import.string -> zone:Core_kernel__.Zone.t -> t

of_filename_string s ~zone converts s that has format YYYY-MM-DD_HH-MM-SS.mmm into time.

val to_string_abs : t -> zone:Core_kernel__.Zone.t -> Core_kernel__.Import.string

to_string_abs ~zone t is the same as to_string t except that it uses the given time zone.

val to_string_abs_trimmed : t -> zone:Core_kernel__.Zone.t -> Core_kernel__.Import.string

to_string_abs_trimmed is the same as to_string_abs, but drops trailing seconds and milliseconds if they are 0.

val to_string_abs_parts : t -> zone:Core_kernel__.Zone.t -> Core_kernel__.Import.string Core_kernel__.Import.list

val to_string_trimmed : t -> zone:Core_kernel__.Zone.t -> Core_kernel__.Import.string

Same as to_string_abs_trimmed, except it leaves off the timezone, so won't reliably round trip.

val to_sec_string : t -> zone:Core_kernel__.Zone.t -> Core_kernel__.Import.string

Same as to_string_abs, but without milliseconds

val of_localized_string : zone:Core_kernel__.Zone.t -> Core_kernel__.Import.string -> t

of_localized_string ~zone str read in the given string assuming that it represents a time in zone and return the appropriate Time.t

val of_string_gen : default_zone:(Core_kernel__.Import.unit -> Core_kernel__.Zone.t) -> find_zone:(Core_kernel__.Import.string -> Core_kernel__.Zone.t) -> Core_kernel__.Import.string -> t

of_string_gen ~default_zone ~find_zone s attempts to parse s as a t, calling out to default_zone and find_zone as needed.

val to_string_iso8601_basic : t -> zone:Core_kernel__.Zone.t -> Core_kernel__.Import.string

to_string_iso8601_basic return a string representation of the following form: %Y-%m-%dT%H:%M:%S.%s%Z e.g. to_string_iso8601_basic ~zone:Time.Zone.utc epoch = "1970-01-01T00:00:00.000000Z"

val occurrence : [ `First_after_or_at | `Last_before_or_at ] -> t -> ofday:Ofday.t -> zone:Core_kernel__.Zone.t -> t

occurrence side time ~ofday ~zone returns a Time.t that is the occurrence of ofday (in the given zone) that is the latest occurrence (<=) time or the earliest occurrence (>=) time, according to side.

NOTE: If the given time converted to wall clock time in the given zone is equal to ofday then the t returned will be equal to the t given.