Module Core_kernel__Union_find

Imperative data structure for representing disjoint sets.

Union find is used to implement an equivalence relation on objects, where the equivalence relation can dynamically be coarsened by "union"ing two equivalence classes together.

All of the operations are effectively (amortized) constant time.

See Wikipedia.

This implementation is not thread-safe.