Module Fheap

Functional heaps (implemented as pairing heaps).

type 'a t

t_of_sexp is not supported, because of the difficulty involved in recreating the comparison function.

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

Even though min_elt, max_elt, and to_list are in Container.S1, they are documented separately to make sure there is no confusion.

include Core_kernel.Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__Container_intf.Export.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

The comparison functions in min_elt and max_elt are independent of the one used to order the heap. Since the provided compare may be different from the one used to create the heap, it is necessary for these functions to traverse the entire heap. If you want to access the smallest element of the heap according to the heap's comparison function, you should use top.

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val to_list : 'a t -> 'a list

The elements of to_list t are not in any particular order. You need to sort the list afterwards if you want to get a sorted list.

val create : cmp:('a -> 'a -> int) -> 'a t

create ~cmp returns a new min-heap that uses ordering function cmp.

The top of the heap is the smallest element as determined by the provided comparison function.

val of_array : 'a array -> cmp:('a -> 'a -> int) -> 'a t

val of_list : 'a list -> cmp:('a -> 'a -> int) -> 'a t

val add : 'a t -> 'a -> 'a t

add t v returns the new heap after addition. Complexity O(1).

val top : 'a t -> 'a option

This returns the top (i.e., smallest) element of the heap. Complexity O(1).

val top_exn : 'a t -> 'a

val remove_top : 'a t -> 'a t option

remove_top t returns the new heap after a remove. It does nothing if t is empty.

The amortized time per remove_top t (or pop t, pop_exn t, pop_if t) is O(lg n). The complexity of the worst case is O(n).

val pop : 'a t -> ('a * 'a t) option

This removes and returns the top (i.e., least) element and the modified heap.

val pop_exn : 'a t -> 'a * 'a t

val pop_if : 'a t -> ('a -> bool) -> ('a * 'a t) option

pop_if t cond returns Some (top_element, rest_of_heap) if t is not empty and its top element satisfies condition cond, or None in any other case.

val to_sequence : 'a t -> 'a Base.Sequence.t

to_sequence t is a sequence of the elements of t in ascending order.