Module Flags__Flags_intf

module Flags implements Unix-style sets of flags that are represented as an int with various bits set, one bit for each flag. E.g. Linux_ext.Epoll.Flag.

Flags defines a module type Flags.S, the interface for a flags, and a functor Flags.Make for creating a flags implementation.