Module Iobuf.Peek

Peek and Poke functions access a value at pos from the lower bound of the window and do not advance.

Peek.bin_prot X.bin_read_t t returns the initial X.t in t without advancing.

Following the bin_prot protocol, the representation of x is X.bin_size_t x bytes long. Peek., Poke., Consume., and Fill.bin_prot do not add any size prefix or other framing to the bin_prot representation.

type 'seek src = (Core_kernel.read, 'seek) t

Similar to Consume.To_*, but do not advance the buffer.

module To_bytes : Core_kernel.Blit.S1_distinct with type 'seek src := 'seek src with type _ dst := Core_kernel.Bytes.t

module To_bigstring : Core_kernel.Blit.S1_distinct with type 'seek src := 'seek src with type _ dst := Core_kernel.Bigstring.t

module To_string : sig ... end

val index : ([> Core_kernel.read ], _) t -> ?pos:int -> ?len:int -> char -> int option

type ('a, 'd, 'w) t = ('d, 'w) t -> pos:int -> 'a

val char : (char, 'd, 'w) t

val int64_t_be : (Core_kernel.Int64.t, 'd, 'w) t

val int64_t_le : (Core_kernel.Int64.t, 'd, 'w) t

val head_padded_fixed_string : padding:char -> len:int -> (string, 'd, 'w) t

val tail_padded_fixed_string : padding:char -> len:int -> (string, 'd, 'w) t

val string : str_pos:int -> len:int -> (string, 'd, 'w) t

val bytes : str_pos:int -> len:int -> (Core_kernel.Bytes.t, 'd, 'w) t

val bigstring : str_pos:int -> len:int -> (Core_kernel.Bigstring.t, 'd, 'w) t

val stringo : ?str_pos:int -> ?len:int -> (string, 'd, 'w) t

val byteso : ?str_pos:int -> ?len:int -> (Core_kernel.Bytes.t, 'd, 'w) t

val bigstringo : ?str_pos:int -> ?len:int -> (Core_kernel.Bigstring.t, 'd, 'w) t

val bin_prot : 'a Core_kernel.Bin_prot.Type_class.reader -> ('a, 'd, 'w) t

val int8 : (int, 'd, 'w) t

val int16_be : (int, 'd, 'w) t

val int16_le : (int, 'd, 'w) t

val int32_be : (int, 'd, 'w) t

val int32_le : (int, 'd, 'w) t

val int64_be_exn : (int, 'd, 'w) t

val int64_le_exn : (int, 'd, 'w) t

val int64_be_trunc : (int, 'd, 'w) t

val int64_le_trunc : (int, 'd, 'w) t

val uint8 : (int, 'd, 'w) t

val uint16_be : (int, 'd, 'w) t

val uint16_le : (int, 'd, 'w) t

val uint32_be : (int, 'd, 'w) t

val uint32_le : (int, 'd, 'w) t

val uint64_be_exn : (int, 'd, 'w) t

val uint64_le_exn : (int, 'd, 'w) t