Module Unsafe.Peek

type 'seek src = ( Core.read, 'seek ) t

Similar to Consume.To_*, but do not advance the buffer.

module To_bytes : Core.Blit.S1_distinct with type 'seek src := 'seek src with type _ dst := Core.Bytes.t

module To_bigstring : Core.Blit.S1_distinct with type 'seek src := 'seek src with type _ dst := Core.Bigstring.t

module To_string : sig ... end

type ('a, 'd, 'w) t = ( 'd, 'w ) t -> pos:int -> 'a

val char : ( char, 'd, 'w ) t

val int64_t_be : ( Core.Int64.t, 'd, 'w ) t

val int64_t_le : ( Core.Int64.t, 'd, 'w ) t

val head_padded_fixed_string : padding:char -> len:int -> ( string, 'd, 'w ) t

val tail_padded_fixed_string : padding:char -> len:int -> ( string, 'd, 'w ) t

val string : str_pos:int -> len:int -> ( string, 'd, 'w ) t

val bytes : str_pos:int -> len:int -> ( Core.Bytes.t, 'd, 'w ) t

val bigstring : str_pos:int -> len:int -> ( Core.Bigstring.t, 'd, 'w ) t

val stringo : ?str_pos:int -> ?len:int -> ( string, 'd, 'w ) t

val byteso : ?str_pos:int -> ?len:int -> ( Core.Bytes.t, 'd, 'w ) t

val bigstringo : ?str_pos:int -> ?len:int -> ( Core.Bigstring.t, 'd, 'w ) t

val bin_prot : 'a Bin_prot.Type_class.reader -> ( 'a, 'd, 'w ) t

val int8 : ( int, 'd, 'w ) t

val int16_be : ( int, 'd, 'w ) t

val int16_le : ( int, 'd, 'w ) t

val int32_be : ( int, 'd, 'w ) t

val int32_le : ( int, 'd, 'w ) t

val int64_be_exn : ( int, 'd, 'w ) t

val int64_le_exn : ( int, 'd, 'w ) t

val int64_be_trunc : ( int, 'd, 'w ) t

val int64_le_trunc : ( int, 'd, 'w ) t

val uint8 : ( int, 'd, 'w ) t

val uint16_be : ( int, 'd, 'w ) t

val uint16_le : ( int, 'd, 'w ) t

val uint32_be : ( int, 'd, 'w ) t

val uint32_le : ( int, 'd, 'w ) t

val uint64_be_exn : ( int, 'd, 'w ) t

val uint64_le_exn : ( int, 'd, 'w ) t

module Int_repr : sig ... end

val index_or_neg : ( [> Core.read ], _ ) iobuf -> pos:int -> len:int -> char -> int

Like Peek.index but with no bounds checks, and returns a negative number rather than None when the character is not found.