Module Unsafe.Poke

Poke.bin_prot X.bin_write_t t x writes x to the beginning of t in binary form without advancing. You can use X.bin_size_t to tell how long it was. X.bin_write_t is only allowed to write that portion of the buffer you have access to.

val decimal : (Core_kernel.read_write, 'w) t -> pos:int -> int -> int

decimal t ~pos i returns the number of bytes written at pos.

val bin_prot_size : 'a Core_kernel.Bin_prot.Type_class.writer -> (Core_kernel.read_write, _) t -> pos:int -> 'a -> int

As bin_prot but returns the number of bytes written.

type ('a, 'd, 'w) t = (Core_kernel.read_write, 'w) t -> pos:int -> 'a -> unit

val char : (char, 'd, 'w) t

val int64_t_be : (Core_kernel.Int64.t, 'd, 'w) t

val int64_t_le : (Core_kernel.Int64.t, 'd, 'w) t

val head_padded_fixed_string : padding:char -> len:int -> (string, 'd, 'w) t

val tail_padded_fixed_string : padding:char -> len:int -> (string, 'd, 'w) t

val string : str_pos:int -> len:int -> (string, 'd, 'w) t

val bytes : str_pos:int -> len:int -> (Core_kernel.Bytes.t, 'd, 'w) t

val bigstring : str_pos:int -> len:int -> (Core_kernel.Bigstring.t, 'd, 'w) t

val stringo : ?str_pos:int -> ?len:int -> (string, 'd, 'w) t

val byteso : ?str_pos:int -> ?len:int -> (Core_kernel.Bytes.t, 'd, 'w) t

val bigstringo : ?str_pos:int -> ?len:int -> (Core_kernel.Bigstring.t, 'd, 'w) t

val bin_prot : 'a Core_kernel.Bin_prot.Type_class.writer -> ('a, 'd, 'w) t

val int8_trunc : (int, 'd, 'w) t

val int16_be_trunc : (int, 'd, 'w) t

val int16_le_trunc : (int, 'd, 'w) t

val int32_be_trunc : (int, 'd, 'w) t

val int32_le_trunc : (int, 'd, 'w) t

val int64_be : (int, 'd, 'w) t

val int64_le : (int, 'd, 'w) t

val uint8_trunc : (int, 'd, 'w) t

val uint16_be_trunc : (int, 'd, 'w) t

val uint16_le_trunc : (int, 'd, 'w) t

val uint32_be_trunc : (int, 'd, 'w) t

val uint32_le_trunc : (int, 'd, 'w) t

val uint64_be_trunc : (int, 'd, 'w) t

val uint64_le_trunc : (int, 'd, 'w) t