Module Nonempty_list

type 'a t =

| :: of 'a * 'a list

A 'a t represents a non-empty list, as evidenced by the fact that there is no [] variant. The sexp representation is as a regular list (i.e., the same as the Stable.V3 module below).

include Ppx_compare_lib.Comparable.S1 with type 'a t := 'a t

val compare : ( 'a -> 'a -> int ) -> 'a t -> 'a t -> int

include Ppx_compare_lib.Equal.S1 with type 'a t := 'a t

val equal : ( 'a -> 'a -> bool ) -> 'a t -> 'a t -> bool

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : ( Sexplib0__.Sexp.t -> 'a ) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ( 'a -> Sexplib0__.Sexp.t ) -> 'a t -> Sexplib0__.Sexp.t

val t_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a t Sexplib0.Sexp_grammar.t

include Ppx_hash_lib.Hashable.S1 with type 'a t := 'a t

val hash_fold_t : ( Base.Hash.state -> 'a -> Base.Hash.state ) -> Base.Hash.state -> 'a t -> Base.Hash.state

include Ppx_quickcheck_runtime.Quickcheckable.S1 with type 'a t := 'a t

val quickcheck_generator : 'a Base_quickcheck.Generator.t -> 'a t Base_quickcheck.Generator.t

val quickcheck_observer : 'a Base_quickcheck.Observer.t -> 'a t Base_quickcheck.Observer.t

val quickcheck_shrinker : 'a Base_quickcheck.Shrinker.t -> 'a t Base_quickcheck.Shrinker.t

include Typerep_lib.Typerepable.S1 with type 'a t := 'a t

val typerep_of_t : 'a Typerep_lib.Std_internal.Typerep.t -> 'a t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : 'a Typerep_lib.Typename.t -> 'a t Typerep_lib.Typename.t

include Core.Comparator.Derived with type 'a t := 'a t

type 'cmp comparator_witness

val comparator : ( 'a, 'cmp ) Core.Comparator.comparator -> ( 'a t, 'cmp comparator_witness ) Core.Comparator.comparator

include Core.Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:( 'a -> 'a -> bool ) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:( 'a -> unit ) -> unit

val fold : 'a t -> init:'accum -> f:( 'accum -> 'a -> 'accum ) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'accum -> f:( 'accum -> 'a -> ( 'accum, 'e ) Base.Result.t ) -> ( 'accum, 'e ) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'accum -> f:( 'accum -> 'a -> ( 'accum, 'final ) Base.Container.Continue_or_stop.t ) -> finish:( 'accum -> 'final ) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:( 'a -> bool ) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:( 'a -> bool ) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:( 'a -> bool ) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base.Container.Summable with type t = 'sum) -> 'a t -> f:( 'a -> 'sum ) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:( 'a -> bool ) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:( 'a -> 'b option ) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:( 'a -> 'a -> int ) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:( 'a -> 'a -> int ) -> 'a option

include Core.Invariant.S1 with type 'a t := 'a t

val invariant : ( 'a -> unit ) -> 'a t -> unit

include Core.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ( 'a -> 'b t ) -> 'b t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ( 'a -> 'b ) -> 'b t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:( 'a -> 'b t ) -> 'b t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:( 'a -> 'b ) -> 'b t

map t ~f is t >>| f.

val join : 'a t t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> unit t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

These are convenient to have in scope when programming with a monad:

val create : 'a -> 'a list -> 'a t

val of_list : 'a list -> 'a t option

val of_list_error : 'a list -> 'a t Core.Or_error.t

val of_list_exn : 'a list -> 'a t

val singleton : 'a -> 'a t

val cons : 'a -> 'a t -> 'a t

val hd : 'a t -> 'a

val tl : 'a t -> 'a list

val reduce : 'a t -> f:( 'a -> 'a -> 'a ) -> 'a

val reverse : 'a t -> 'a t

val append : 'a t -> 'a list -> 'a t

val unzip : ('a * 'b) t -> 'a t * 'b t

val zip : 'a t -> 'b t -> ('a * 'b) t Core.List.Or_unequal_lengths.t

val zip_exn : 'a t -> 'b t -> ('a * 'b) t

val mapi : 'a t -> f:( int -> 'a -> 'b ) -> 'b t

val concat : 'a t t -> 'a t

val concat_map : 'a t -> f:( 'a -> 'b t ) -> 'b t

val last : 'a t -> 'a

val drop_last : 'a t -> 'a list

val to_sequence : 'a t -> 'a Core.Sequence.t

val sort : 'a t -> compare:( 'a -> 'a -> int ) -> 'a t

val stable_sort : 'a t -> compare:( 'a -> 'a -> int ) -> 'a t

val dedup_and_sort : compare:( 'a -> 'a -> int ) -> 'a t -> 'a t

val fold_right : 'a t -> init:'b -> f:( 'a -> 'b -> 'b ) -> 'b

val folding_map : 'a t -> init:'b -> f:( 'b -> 'a -> 'b * 'c ) -> 'c t

val min_elt' : 'a t -> compare:( 'a -> 'a -> int ) -> 'a

min_elt' and max_elt' differ from min_elt and max_elt (included in Container.S1) in that they don't return options.

val max_elt' : 'a t -> compare:( 'a -> 'a -> int ) -> 'a

val map_of_alist_multi : ('k * 'v) list -> comparator:( 'k, 'cmp ) Core.Map.comparator -> ( 'k, 'v t, 'cmp ) Core.Map.t

Like Map.of_alist_multi, but comes with a guarantee that the range of the returned map is all nonempty lists.

val map_of_sequence_multi : ('k * 'v) Core.Sequence.t -> comparator:( 'k, 'cmp ) Core.Map.comparator -> ( 'k, 'v t, 'cmp ) Core.Map.t

Like Map.of_sequence_multi, but comes with a guarantee that the range of the returned map is all nonempty lists.

val combine_errors : ( 'ok, 'err ) Core.Result.t t -> ( 'ok t, 'err t ) Core.Result.t

Like Result.combine_errors but for non-empty lists

val combine_errors_unit : ( unit, 'err ) Core.Result.t t -> ( unit, 'err t ) Core.Result.t

Like Result.combine_errors_unit but for non-empty lists

val validate_indexed : 'a Core.Validate.check -> 'a t Core.Validate.check

validates a list, naming each element by its position in the list (where the first position is 1, not 0).

val validate : name:( 'a -> string ) -> 'a Core.Validate.check -> 'a t Core.Validate.check

validates a list, naming each element using a user-defined function for computing the name.

val flag : 'a Core.Command.Param.Arg_type.t -> 'a t Core.Command.Flag.t

type 'a nonempty_list := 'a t

module Option : sig ... end

This module provides 0-alloc versions of to_list and of_list, via some and allowing you to match%optional on a list, respectively.

module Reversed : sig ... end

val rev_append : 'a Reversed_list.t -> 'a t -> 'a t

module Unstable : sig ... end

module Stable : sig ... end