Module Nonempty_list.Stable

module V3 : sig ... end

Represents a t as an ordinary list for sexp and bin_io conversions, e.g. 1::2 is represented as (1 2).

module V2 : sig ... end

Represents a t as an ordinary list for sexp conversions, but uses a record {hd : 'a; tl ; 'a list} for bin_io conversions.

module V1 : sig ... end

Represents a t as an ordinary list for sexps, but as a pair for bin_io conversions (i.e., a 'a t is represented as the type 'a * 'a list).