Module Pairing_heap

Heap implementation based on a pairing-heap.

This heap implementations supports an arbitrary element type via a comparison function.

type 'a t

of_sexp and bin_io functions aren't supplied for heaps due to the difficulties in reconstructing the correct comparison function when de-serializing.

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

Mutation of the heap during iteration is not supported, but there is no check to prevent it. The behavior of a heap that is mutated during iteration is undefined.

include Core_kernel.Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Base.Result.t) -> ('accum, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__Container_intf.Export.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

include Core_kernel.Invariant.S1 with type 'a t := 'a t

val invariant : ('a -> unit) -> 'a t -> unit

Even though these two functions min_elt and max_elt are part of Container.S1, they are documented separately to make sure there is no confusion. They are independent of the comparison function used to order the heap. Instead, a traversal of the entire structure is done using the provided cmp function to find a min or max.

If you want to access the smallest element of the heap according to the heap's comparison function in constant time, you should use top.

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val create : ?min_size:int -> cmp:('a -> 'a -> int) -> unit -> 'a t

create ?min_size ~cmp returns a new min-heap that can store min_size elements without reallocations, using ordering function cmp.

The top of the heap is the smallest element as determined by the provided comparison function. In particular, if cmp x y < 0 then x will be "on top of" y in the heap.

Memory use can be surprising in that the underlying pool never shrinks, so current memory use will at least be proportional to the largest number of elements that the heap has ever held.

val of_array : 'a array -> cmp:('a -> 'a -> int) -> 'a t

min_size (see create) will be set to the size of the input array or list.

val of_list : 'a list -> cmp:('a -> 'a -> int) -> 'a t

val top : 'a t -> 'a option

Returns the top (i.e., smallest) element of the heap.

val top_exn : 'a t -> 'a

val add : 'a t -> 'a -> unit

val remove_top : _ t -> unit

remove_top t does nothing if t is empty.

val pop : 'a t -> 'a option

pop removes and returns the top (i.e. least) element.

val pop_exn : 'a t -> 'a

val pop_if : 'a t -> ('a -> bool) -> 'a option

pop_if t cond returns Some top_element of t if it satisfies condition cond, removing it, or None in any other case.

val copy : 'a t -> 'a t

copy t returns a shallow copy.

module Elt : sig ... end

val add_removable : 'a t -> 'a -> 'a Elt.t

add_removable t v adds v to t, returning a token that can be used to delete v from t in lg(n) amortized time.

Note that while add doesn't allocate unless the underlying pool needs to be resized, add_removable always allocates. The Unsafe module has a non-allocating alternative.

val remove : 'a t -> 'a Elt.t -> unit

If t and token are mismatched then behavior is undefined. Trying to remove an already removed token (by an earlier call to remove or pop for instance) is a no-op, but keeping token around after it has been removed may lead to memory leaks since it has a reference to the heap.

val update : 'a t -> 'a Elt.t -> 'a -> 'a Elt.t

update t token v is shorthand for remove t token; add_removable t v.

val find_elt : 'a t -> f:('a -> bool) -> 'a Elt.t option

find_elt t ~f. If f is true for some element in t, return an Elt.t for that element. This operation is O(n).

module Unsafe : sig ... end

Unsafe functions provide faster alternatives to regular functions with the same name. They don't allocate but the behavior is unspecified and could be memory unsafe in certain cases where regular functions would fail with informative exceptions.