Module Pairing_heap.Elt

type 'a t

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

val value_exn : 'a t -> 'a

value_exn t returns the value in the heap controlled by this token if the value is still in the heap, and raises otherwise.