Module Sexp_hidden_in_test

type 'a t = 'a

'a t is a type that uses Core_kernel.am_running_test to determine if it should use the 'a sexp serializer, or serialize the type as '<hidden_in_test>'. It can be thought of as a form of @sexp.opaque that is conditional upon if tests are running.

include Core_kernel.Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val sexp_of_t : ('a -> Ppx_sexp_conv_lib.Sexp.t) -> 'a t -> Ppx_sexp_conv_lib.Sexp.t

module With_non_roundtripping_in_test_of_sexp : sig ... end

This type also derives sexp. This will not allow you to roundtrip values you create in tests and should be used carefully.