Module Interval_num.Span

type t = private Core_kernel.Int63.t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

include Core_kernel.Comparable.S with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Replace_polymorphic_compare : Base.Comparable.Polymorphic_compare with type t := t

module Map : Core_kernel.Map.S with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core_kernel.Set.S with type Elt.t = t with type Elt.comparator_witness = comparator_witness

val max : t -> t -> t

val zero : t

val one : t

val max_value : t

val of_int63 : Core_kernel.Int63.t -> t

val to_int63 : t -> Core_kernel.Int63.t

val of_int : int -> t

val to_int_exn : t -> int

val scale_int : t -> int -> t

val pred : t -> t

val succ : t -> t

val (+) : t -> t -> t