Module S.Pointer

type 'slots t

A pointer to a tuple in a pool. 'slots will look like ('a1, ..., 'an) Slots.tn, and the tuples have type 'a1 * ... * 'an.

val sexp_of_t : ('slots -> Ppx_sexp_conv_lib.Sexp.t) -> 'slots t -> Ppx_sexp_conv_lib.Sexp.t

include Typerep_lib.Typerepable.S1 with type 'slots t := 'slots t

val typerep_of_t : 'a Typerep_lib.Std_internal.Typerep.t -> 'a t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : 'a Typerep_lib.Typename.t -> 'a t Typerep_lib.Typename.t

val null : unit -> _ t

The null pointer is a distinct pointer that does not correspond to a tuple in the pool. It is a function to prevent problems due to the value restriction.

val is_null : _ t -> bool

val phys_compare : 'a t -> 'a t -> int

val phys_equal : 'a t -> 'a t -> bool

module Id : sig ... end