Module Unpack_buffer.Unpack_one

If unpack_one : ('a, 'state) unpack, then unpack_one ~state ~buf ~pos ~len must unpack at most one value of type 'a from buf starting at pos, and not using more than len characters. unpack_one must return one the following:

`Ok (value, n) -- unpacking succeeded and consumed n bytes, where 0 <= n <= len. It is possible to have n = 0, e.g. for sexp unpacking, which can only tell it has reached the end of an atom when it encounters the following punctuation character, which if it is left paren, is the start of the following sexp.

`Not_enough_data (state, n) -- unpacking encountered a valid proper prefix of a packed value, and consumed n bytes, where 0 <= n <= len. state can be supplied to a future call to unpack_one to continue unpacking.

`Invalid_data -- unpacking encountered an invalidly packed value.

A naive unpack_one that only succeeds on a fully packed value could lead to quadratic behavior if a packed value's bytes are input using a linear number of calls to feed.

type ('a, 'state) unpack_result = [

| `Ok of 'a * int

| `Not_enough_data of 'state * int

| `Invalid_data of Core_kernel.Error.t

]

type ('a, 'state) unpack = state:'state -> buf:Core_kernel.Bigstring.t -> pos:int -> len:int -> ('a, 'state) unpack_result

type 'a t =

| T : {

initial_state : 'state;

unpack : ('a, 'state) unpack;

} -> 'a t

include Core_kernel.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:('a -> 'b) -> 'b t

map t ~f is t >>| f.

val join : 'a t t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> unit t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

These are convenient to have in scope when programming with a monad:

val create : initial_state:'state -> unpack:('a, 'state) unpack -> 'a t

val create_bin_prot : 'a Core_kernel.Bin_prot.Type_class.reader -> 'a t

create_bin_prot reader returns an unpacker that reads the "size-prefixed" bin_prot encoding, in which a value is encoded by first writing the length of the bin_prot data as a 64-bit int, and then writing the data itself.

val bin_blob : Core_kernel.Bin_prot.Blob.Opaque.Bigstring.t t

Reads "size-prefixed" bin-blobs, much like create_bin_prot _, but preserves the size information and doesn't deserialize the blob. This allows deserialization to be deferred and the remainder of the sequence can be unpacked if an individual blob can't be deserialized.

val sexp : Core_kernel.Sexp.t t

Beware that when unpacking sexps, one cannot tell if one is at the end of an atom until one hits punctuation. So, one should always feed a space (" ") to a sexp unpack buffer after feeding a batch of complete sexps, to ensure that the final sexp is unpacked.

val char : char t

module type Equal = sig ... end

val expect : 'a t -> (module Equal with type t = 'a) -> 'a -> unit t

expect t equal a returns an unpacker that unpacks using t and then returns `Ok if the unpacked value equals a, or `Invalid_data otherwise.

val expect_char : char -> unit t

expect_char is expect char (module Char)

val newline : unit t