Module Uuid

Implements universally unique identifiers based on version 3 of the UUID specification. Identifier generation is thread safe, and fast.

type t

When am_running_test, sexp_of_t shows all zeros (the nil UUID).

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

include Core_kernel.Identifiable.S with type t := t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Core_kernel.Identifiable.S_common with type t := t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Base.Pretty_printer.S with type t := t

val pp : Base.Formatter.t -> t -> unit

include Core_kernel.Comparable.S_binable with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Replace_polymorphic_compare : Base.Comparable.Polymorphic_compare with type t := t

include Core_kernel.Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : (t, comparator_witness) Core_kernel.Comparator.comparator

module Map : Core_kernel.Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core_kernel.Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Core_kernel.Hashable.S_binable with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Core_kernel.Hashtbl.Hashable.t

module Table : Core_kernel.Hashtbl.S_binable with type key = t

module Hash_set : Core_kernel.Hash_set.S_binable with type elt = t

module Hash_queue : Core_kernel.Hash_queue.S with type key = t

include Core_kernel.Invariant.S with type t := t

val invariant : t -> unit

include Core_kernel.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val t_of_sexp : Core_kernel.Sexp.t -> t

val create : unit -> t

val create_random : Core_kernel.Random.State.t -> t

val to_string_hum : t -> string

to_string_hum t is like to_string, except when am_running_test, in which case it shows all zeros (the nil UUID).

val arg_type : t Core_kernel.Command.Arg_type.t

module Unstable : sig ... end

module Stable : sig ... end