Module Uuid.Hash_set

type elt = t

type t = elt Core_kernel.Hash_set.t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

type 'a t_ = t

type 'a elt_ = elt

val create : ?growth_allowed:bool -> ?size:int -> unit -> 'a t_

val of_list : ?growth_allowed:bool -> ?size:int -> 'a elt_ list -> 'a t_

module Provide_of_sexp (X : sig ... end) : sig ... end

module Provide_bin_io (X : sig ... end) : sig ... end

val t_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t