Module Version_util.Expert

val insert_version_util : contents_of_exe:string -> Version.t list -> string option

Inserts the given version util into the executable text given. Returns None if this could not happen (maybe this is an executable that doesn't link in the current library, maybe it already has version util).

val remove_version_util : contents_of_exe:string -> string option

Undoes the operation above. Note that as currently implemented, this only removes version util added by insert_version_util. It does not remove version added when building with VERSION_UTIL_SUPPORT=true. The point of this operation is to make it possible to compare executables up to version util.

module For_tests : sig ... end