Module Bounded_int_table.With_key

Parameters

module Key : sig ... end

Signature

type 'data t = ( Key.t, 'data ) table

Serialization of a bounded int table using bin_io or sexp preserves num_keys, but only takes space proportional to the length of the table.

include Bin_prot.Binable.S1 with type 'data t := 'data t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ( 'a, 'a t ) Bin_prot.Size.sizer1

val bin_write_t : ( 'a, 'a t ) Bin_prot.Write.writer1

val bin_read_t : ( 'a, 'a t ) Bin_prot.Read.reader1

val __bin_read_t__ : ( 'a, int -> 'a t ) Bin_prot.Read.reader1

val bin_writer_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.writer

val bin_reader_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.reader

val bin_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.t

include Sexplib0.Sexpable.S1 with type 'data t := 'data t

val t_of_sexp : ( Sexplib0__.Sexp.t -> 'a ) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ( 'a -> Sexplib0__.Sexp.t ) -> 'a t -> Sexplib0__.Sexp.t

val create : num_keys:int -> 'data t

val of_alist : (Key.t * 'data) list -> 'data t Core.Or_error.t

of_alist returns a table whose maximum allowed key is the maximum key in the input list.

val of_alist_exn : (Key.t * 'data) list -> 'data t