Module Hash_heap

A hash-heap is a combination of a heap and a hashtable that supports constant time lookup, and log(n) time removal and replacement of elements in addition to the normal heap operations.

module Make : functor (Key : Hash_heap__.Hash_heap_intf.Key) -> Hash_heap__.Hash_heap_intf.S with module Key = Key