Module Hash_heap

A hash-heap is a combination of a heap and a hashtable that supports constant time lookup, and log(n) time removal and replacement of elements in addition to the normal heap operations.

module Make (Key : Core.Hashtbl.Key_plain) : sig ... end